% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AL14B1.R
\name{AL14B1}
\alias{AL14B1}
\title{Alroy's (2014) "Creeping Shadow of a Doubt" model}
\usage{
AL14B1(records)
}
\arguments{
\item{records}{sighting records in \code{cbin} format (see
\code{\link{convert_dodo}} for details).}
}
\value{
a \code{list} object with the original parameters and p(extant) included
as elements.
}
\description{
The model from Alroy 2014. Estimates a posterior probability that the
species is extant at the end of the observation period.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\examples{
# Run an example analysis using the Caribbean Monk Seal data
AL14B1(as.integer(1915:1992 \%in\% monk_seal))
# Run an example analysis using the Slender-billed Curlew data
\dontrun{
AL14B1(curlew$cbin)
}

}
\references{
\strong{Key Reference}

Alroy, J. (2014). A simple Bayesian method of inferring extinction.
\emph{Paleobiology}, 40(4), 584-607. \doi{10.1666/13074}

\strong{Other References}

Alroy, J. (2016). On a conservative Bayesian method of inferring extinction.
\emph{Paleobiology}, 42(4), 670-679. \doi{10.1017/pab.2016.12}

Alroy, J. (2016). A simple Bayesian method of inferring extinction: reply.
\emph{Ecology}, 97(3), 798-800. \doi{10.1002/ecy.1321}
}
\seealso{
\code{\link[=AL15B1]{AL15B1()}}]
}
