% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AL15B1.R
\name{AL15B1}
\alias{AL15B1}
\title{Alroy's (2015) "Agnostic" model}
\usage{
AL15B1(records)
}
\arguments{
\item{records}{sighting records in \code{cbin} format (see
\code{\link{convert_dodo}} for details).}
}
\value{
a \code{list} object with the original parameters and p(extant) included
as elements.
}
\description{
The model from Alroy 2015. Estimates a posterior probability that the
species is extant at the end of the observation period.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant. Uses a uniform prior, as suggested by Alroy (pers. comm.).
}
\examples{
# Run an example analysis using the Caribbean Monk Seal data
AL15B1(as.integer(1915:1992 \%in\% monk_seal))
# Run an example analysis using the Slender-billed Curlew data
\dontrun{
AL15B1(curlew$cbin)
}

}
\references{
\strong{Key Reference}

Alroy, J. (2015). Current extinction rates of reptiles and amphibians.
\emph{Proceedings of the National Academy of Sciences of the United States of}
\emph{America}, 112(42), 13003-13008. \doi{10.1073/pnas.1508681112}

\strong{Other References}

Alroy, J. (2016). On a conservative Bayesian method of inferring extinction.
\emph{Paleobiology}, 42(4), 670-679. \doi{10.1017/pab.2016.12}

Alroy, J. (2016). A simple Bayesian method of inferring extinction: reply.
\emph{Ecology}, 97(3), 798-800. \doi{10.1002/ecy.1321}
}
\seealso{
\code{\link[=AL14B1]{AL14B1()}}]
}
