% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anabarella}
\alias{anabarella}
\title{\emph{Anabarella} fossil record}
\format{
A \code{numeric} vector with 19 values.
}
\usage{
anabarella
}
\description{
Siberian and Mongolian records of Cambrian \emph{Anabarella} sp. from Wang et al.
(2016) based on data presented in Maloof et al. (2010). Format: \code{ccon}.
}
\note{
Dates are measured in million years before present (note that values are all
<0). Radiometric dating error estimates are not included.
}
\references{
\strong{Key Reference}

Wang, S. C., Everson, P. J., Zhou, H. J., Park, D., & Chudzicki, D. J.
(2016). Adaptive credible intervals on stratigraphic ranges when recovery
potential is unknown. \emph{Paleobiology}, 42(2), 240-256.
\doi{10.1017/pab.2015.37}

\strong{Other Reference}

Maloof, A. C., Porter, S. M., Moore, J. L., Dudás, F. Ö., Bowring, S. A.,
Higgins, J. A., ... & Eddy, M. P. (2010). The earliest Cambrian record of
animals and ocean geochemical change. \emph{GSA Bulletin}, 122(11-12), 1731-1774.
\doi{10.1130/B30346.1}
}
\keyword{datasets}
