% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-utilities.R
\name{benchmark_convert_concept}
\alias{benchmark_convert_concept}
\title{Benchmarking: Convert concept identifiers using benchmark edges}
\usage{
benchmark_convert_concept(
  from,
  to = NULL,
  from.concept = c("Disease", "Phenotype"),
  to.concept = c("Disease", "Phenotype"),
  deprecated = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{from}{a vector with identifier to convert formatted as DB:id (eg. "MONDO:0005027)}

\item{to}{database to convert to (default = NULL, no filtering)}

\item{from.concept}{concept (disease or phenotype) of from}

\item{to.concept}{concept (disease or phenotype) to convert to}

\item{deprecated}{include deprecated identifiers (default: false)}

\item{verbose}{show query input (default: FALSE)}
}
\value{
a dataframe with three columns:
\itemize{
\item from: identifier to convert
\item to: returned conversion
}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
\details{
To assess the different cutoffs used to define \emph{is_xref} and \emph{is_related} edges
the convert_concept function is used. As additional edges are defined (_bm) for testing
this function uses these edges to compare conversion using different BA cutoffs
}
