% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-utilities.R
\name{benchmark_remove_edges}
\alias{benchmark_remove_edges}
\title{Removing benchmark edges}
\usage{
benchmark_remove_edges(type, concept)
}
\arguments{
\item{type}{edge type}

\item{concept}{either "Disease" or "Phenotype"}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
