% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bitterweed}
\alias{bitterweed}
\title{Bitterweed sightings}
\format{
a \code{numeric} vector with 178 values.
}
\usage{
bitterweed
}
\description{
Bitterweed (\emph{Helenium amarum}) sighting record from Rout et al. (2009).
Format: \code{cbin}.
}
\references{
\strong{Key Reference}

Rout, T. M., Salomon, Y., & McCarthy, M. A. (2009). Using sighting records
to declare eradication of an invasive species. \emph{Journal of Applied Ecology},
46(1), 110-117.\doi{10.1111/j.1365-2664.2008.01586.x}
}
\keyword{datasets}
