% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BR12F1.R
\name{BR12F1}
\alias{BR12F1}
\title{Bradshaw et al.'s (2012) "Inverse-weighted McInerny" model}
\usage{
BR12F1(records, alpha = 0.05)
}
\arguments{
\item{records}{sighting records in \code{ccon} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) to
use in the underlying McInerny et al. 2006 model.}
}
\value{
a \code{list} object with the original parameters and the point estimate
included as elements.
}
\description{
Equation 4 from Bradshaw et al. 2012. Estimates a point estimate on the time
of extinction.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant. This is \emph{not} the Gaussian-resampled model (GRIWM) also
presented in Bradshaw et al. 2012 (which incorporates radiometric dating
error). Duplicate sightings in a single time unit are discarded.
}
\examples{
# Run the Woolly Mammoth analysis from Bradshaw et al. 2012
BR12F1(mammoth)
\dontrun{
# Run an example analysis using the Slender-billed Curlew data
BR12F1(curlew$ccon)
}

}
\references{
\strong{Key Reference}

Bradshaw, C. J. A., Cooper, A., Turney, C. S. M., & Brook, B. W. (2012).
Robust estimates of extinction time in the geological record. \emph{Quaternary}
\emph{Science Reviews}, 33, 14-19. \doi{10.1016/j.quascirev.2011.11.021}

\strong{Other References}

McInerny, G. J., Roberts, D. L., Davy, A. J., & Cribb, P. J. (2006).
Significance of sighting rate in inferring extinction and threat.
\emph{Conservation Biology}, 20(2), 562-567.
\doi{10.1111/j.1523-1739.2006.00377.x}
}
