% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BR19F2.R
\name{BR19F2}
\alias{BR19F2}
\title{Brook et al.'s (2019) "Solow & Roberts" model}
\usage{
BR19F2(
  records,
  alpha = 0.05,
  test.time = as.numeric(format(Sys.Date(), "\%Y")),
  cores = NULL,
  n.iter = 10000
)
}
\arguments{
\item{records}{sighting records in \code{ucon} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}

\item{cores}{number of cores to use (defaults to \code{NULL}, in which case
\code{parallel::detectCores()} is run).}

\item{n.iter}{number of iterations to run (defaults to 10000).}
}
\value{
a \code{list} object with the original parameters and the probability,
point estimate, and confidence interval included as elements. The confidence
interval is a two-element numeric vector called \code{conf.int}.
}
\description{
The Solow & Roberts 2003 variant of the model from Brook et al. 2019.
Estimates a "probability" of extinction at the test time, and a two-tailed
\eqn{1 - \alpha} confidence interval and point estimate on the time of
extinction. Sighting uncertainty is incorporated.
}
\note{
Sampling effort is assumed to be constant.
}
\examples{
# Run the "Extreme" Ivory-billed Woodpecker analysis from Brook et al. 2019
BR19F2(woodpecker$ucon, test.time = 2010, alpha = 0.1, cores = 2,
       n.iter = 1e3)
# NB: alpha = 0.1 as this package presents two-sided confidence intervals
# vs. one-sided confidence intervals in the original paper; the upper bound
# of the one-sided alpha = 0.05 CI in Brook et al. is equivalent to the
# upper bound of the two-sided alpha = 0.1 CI in this example.
# Run an example analysis using the Slender-billed Curlew data
BR19F2(curlew$ucon, test.time = 2022, cores = 2, n.iter = 1e3)

}
\references{
\strong{Key Reference}

Brook, B. W., Buettel, J. C., & Jaric, I. (2019). A fast re-sampling method
for using reliability ratings of sightings with extinction-date estimators.
\emph{Ecology}, 100(9), e02787. \doi{10.1002/ecy.2787}

\strong{Other References}

Solow, A. R., & Roberts, D. L. (2003). A nonparametric test for extinction
based on a sighting record. \emph{Ecology}, 84(5), 1329-1332.
\doi{10.1890/0012-9658(2003)084[1329:ANTFEB]2.0.CO;2}

Robson, D. S., & Whitlock, J. H. (1964). Estimation of a truncation point.
\emph{Biometrika}, 51(1-2), 33-39. \doi{10.1093/biomet/51.1-2.33}
}
\seealso{
\code{\link[=BR19F1]{BR19F1()}}, \code{\link[=BR19F3]{BR19F3()}}, \code{\link[=BR19F4]{BR19F4()}}, \code{\link[=BR19F5]{BR19F5()}}
}
