% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BR19F4.R
\name{BR19F4}
\alias{BR19F4}
\title{Brook et al.'s (2019) "McInerny" model}
\usage{
BR19F4(
  records,
  alpha = 0.05,
  test.time = as.numeric(format(Sys.Date(), "\%Y")),
  cores = NULL,
  n.iter = 10000
)
}
\arguments{
\item{records}{sighting records in \code{ucon} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}

\item{cores}{number of cores to use (defaults to \code{NULL}, in which case
\code{parallel::detectCores()} is run).}

\item{n.iter}{number of iterations to run (defaults to 10000).}
}
\value{
a \code{list} object with the original parameters and the probability,
point estimate, and confidence interval included as elements. The confidence
interval is a two-element numeric vector called \code{conf.int}.
}
\description{
The McInerny et al. 2006 variant of the model from Brook et al. 2019.
Estimates a "probability" of extinction at the test time, and a two-tailed
\eqn{1 - \alpha} confidence interval and point estimate on the time of
extinction. Sighting uncertainty is incorporated.
}
\note{
Sampling effort is assumed to be constant.
}
\examples{
# Run the "Extreme" Ivory-billed Woodpecker analysis from Brook et al. 2019
BR19F4(woodpecker$ucon, test.time = 2010, alpha = 0.1, cores = 2,
       n.iter = 1e3)
# NB: alpha = 0.1 as this package presents two-sided confidence intervals
# vs. one-sided confidence intervals in the original paper; the upper bound
# of the one-sided alpha = 0.05 CI in Brook et al. is equivalent to the
# upper bound of the two-sided alpha = 0.1 CI in this example.
# Run an example analysis using the Slender-billed Curlew data
BR19F4(curlew$ucon, test.time = 2022, cores = 2, n.iter = 1e3)

}
\references{
\strong{Key Reference}

Brook, B. W., Buettel, J. C., & Jaric, I. (2019). A fast re-sampling method
for using reliability ratings of sightings with extinction-date estimators.
\emph{Ecology}, 100(9), e02787. \doi{10.1002/ecy.2787}

\strong{Other References}

McInerny, G. J., Roberts, D. L., Davy, A. J., & Cribb, P. J. (2006).
Significance of sighting rate in inferring extinction and threat.
\emph{Conservation Biology}, 20(2), 562-567.
\doi{10.1111/j.1523-1739.2006.00377.x}
}
\seealso{
\code{\link[=BR19F1]{BR19F1()}}, \code{\link[=BR19F2]{BR19F2()}}, \code{\link[=BR19F3]{BR19F3()}}, \code{\link[=BR19F5]{BR19F5()}}
}
