% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BU95F1.R
\name{BU95F1}
\alias{BU95F1}
\title{Burgman et al.'s (1995) "Discrete-time" model}
\usage{
BU95F1(records)
}
\arguments{
\item{records}{\code{data.frame} with two columns: \code{time} and \code{records}. The
\code{time} column must extend from the start of the observation period (which
may be prior to the first sighting) to the end (typically the present day),
with evenly-spaced temporal intervals (typically years).}
}
\value{
a \code{list} object with the original parameters and the p-value
included as elements.
}
\description{
Equation 2 from Burgman et al. 1995. Estimates a p-value for
testing competing hypotheses of extinction/non-extinction.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\examples{
# Run the example analysis from Burgman 1995 (Figure 1b)
BU95F1(burgman_figure1b)

}
\references{
\strong{Key Reference}

Burgman, M. A., Grimson, R. C., & Ferson, S. (1995). Inferring Threat from
Scientific Collections. \emph{Conservation Biology}, 9(4), 923-928.
\doi{0.1046/j.1523-1739.1995.09040923.x}
}
\seealso{
\code{\link[=BU95F2]{BU95F2()}}, \code{\link[=BU95F3]{BU95F3()}}
}
