% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BU95F3.R
\name{BU95F3}
\alias{BU95F3}
\title{Burgman et al.'s (1995) "Empty Cells" model}
\usage{
BU95F3(records)
}
\arguments{
\item{records}{sighting records in \code{cdis} format (see
\code{\link{convert_dodo}} for details).}
}
\value{
a \code{list} object with the original parameters and the p-value
included as elements.
}
\description{
Equation 4 from Burgman et al. 1995. Estimates a p-value for
testing competing hypotheses of extinction/non-extinction.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.

Uses code from the \code{sExtinct} package
(\href{https://github.com/ExperimentalConservation/sExtinct}{GitHub})
}
\examples{
# Run the example analysis from Burgman 1995 (Figure 1b)
BU95F3(burgman_figure1b)
# Run an example analysis using the Slender-billed Curlew data
\dontrun{BU95F3(curlew$cdis)}

}
\references{
\strong{Key Reference}

Burgman, M. A., Grimson, R. C., & Ferson, S. (1995). Inferring Threat from
Scientific Collections. \emph{Conservation Biology}, 9(4), 923-928.
\doi{0.1046/j.1523-1739.1995.09040923.x}

\strong{Other References}

Clements, C. F., Collen, B., Blackburn, T. M., & Petchey, O. L. (2014).
Effects of recent environmental change on accuracy of inferences of
extinction status. \emph{Conservation Biology}, 28(4), 971-981.
\doi{10.1111/cobi.12329}

Grimson, R. C., Aldrich, T. E., & Drane, J. W. (1992). Clustering in sparse
data and an analysis of rhabdomyosarcoma incidence.
\emph{Statistics in Medicine}, 11(6), 761-768. \doi{10.1002/sim.4780110607}
}
\seealso{
\code{\link[=BU95F1]{BU95F1()}}, \code{\link[=BU95F2]{BU95F2()}}
}
