% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{build_multicypher}
\alias{build_multicypher}
\title{Helper to build queries for multicypher}
\usage{
build_multicypher(statements, result = "row", parameters = NULL)
}
\arguments{
\item{statements}{cypher query}

\item{result}{the way to return results. "row" will return a data frame
and "graph" will return a list of nodes, a list of relationships and
a list of paths (vectors of relationships identifiers).}

\item{parameters}{parameters for the cypher query}
}
\description{
Helper to build queries for multicypher
}
