% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDisNet.R
\name{buildDisNetByID}
\alias{buildDisNetByID}
\title{Build a network of diseases according to provided identifiers}
\usage{
buildDisNetByID(
  ids,
  backwardAmbiguity = 1,
  forwardAmbiguity = NULL,
  avoidOrigin = NULL
)
}
\arguments{
\item{ids}{a character vector of identifier to search formatted as \emph{DB:id} (e.g. "MONDO:0005027")}

\item{backwardAmbiguity}{level of backward ambiguity allowed
(default: 1 ==> no ambiguity allowed)}

\item{forwardAmbiguity}{level of forward ambiguity allowed
(default: 10000 ==> ~no filter)}

\item{avoidOrigin:}{allows to avoid traversing parent/child edges from a particular ontology (default = NULL)}
}
\value{
A normalized disease network
}
\description{
Querying for (a) disease identifier(s) is simplified using the
helper function \emph{buildDisNetByID} using a vector or dataframe as input.
}
\details{
The disNet object is constructed around the return
query results also providing information on their relationships. Therefore the \emph{buildDisNet} function
doesn't deal with transitivity of cross-reference edges, so the forward and backward ambiguity
are applied for both type of cross-references (\emph{is_xref} and \emph{is_related}) consistently.
The default implementation only has a limit on the backward ambiguity (see vignette).
In addition, it is possible to avoid returning a particular
hierarchical relationship by specifiying \emph{avoidOrigin}.
}
\examples{

buildDisNetByID(ids = c("ORPHA:86814","MONDO:0005099"))

}
\seealso{
buildDisNet, buildDisNetByTerm
}
