% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burgman_figure1b}
\alias{burgman_figure1b}
\title{Example sightings from Burgman et al.}
\format{
a \code{numeric} vector with 16 values.
}
\usage{
burgman_figure1b
}
\description{
Example frequency-based sighting record from Figure 1b in Burgman et al.
(1995). The temporal range (\code{init.time} to \code{test.time}) of these data is 0 to
16. Format: \code{cdis}.
}
\references{
\strong{Key Reference}

Burgman, M. A., Grimson, R. C., & Ferson, S. (1995). Inferring Threat from
Scientific Collections. Conservation Biology, 9(4), 923-928.
\doi{0.1046/j.1523-1739.1995.09040923.x}
}
\keyword{datasets}
