% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burgmanf1b}
\alias{burgmanf1b}
\title{Example sightings from Burgman et al.}
\format{
a \code{data.frame} with 2 columns and 16 rows.
}
\usage{
burgmanf1b
}
\description{
Example frequency-based sighting records from Figure 1b in Burgman et al.
(1995).
}
\references{
\strong{Key Reference}
Burgman, M. A., Grimson, R. C., & Ferson, S. (1995). Inferring Threat from
Scientific Collections. Conservation Biology, 9(4), 923-928.
\doi{0.1046/j.1523-1739.1995.09040923.x}
}
\keyword{datasets}
