% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_disNet.R
\name{c.disNet}
\alias{c.disNet}
\title{Merge several disease networks}
\usage{
\method{c}{disNet}(..., list = NULL)
}
\arguments{
\item{...}{disease networks to merge}

\item{list}{list object with disease networks to merge}
}
\value{
A normalized merged disease network
}
\description{
Merge several disease networks
}
\examples{
disNet1 <- build_disNet(term = "epilep")
disNet2 <- build_disNet(term = "alzheim")
mergeDisNet <- c(disNet1, disNet2)

}
