% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_dodo.R
\name{cache_dodo_results}
\alias{cache_dodo_results}
\title{Put a DODO query result in cache}
\usage{
cache_dodo_results(value, name)
}
\arguments{
\item{value}{the result to cache}

\item{name}{the name of the query}
}
\description{
Internal use
}
\seealso{
\code{\link[=call_dodo_cache]{call_dodo_cache()}}, \code{\link[=load_dodo_results]{load_dodo_results()}}
}
