% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-utilities.R
\name{calculate_ambiguity}
\alias{calculate_ambiguity}
\title{Ambiguity calculations}
\usage{
calculate_ambiguity(DODO_crossId)
}
\arguments{
\item{DODO_crossId}{object with cross-reference information

Not exported to avoid unintended modifications of the DB.}
}
\value{
object with cross-reference information and added ambiguity edges
}
\description{
Calculating the forward and backward ambiguity on cross-reference edges
}
