% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{call_dodo}
\alias{call_dodo}
\title{Call a function on the DODO graph}
\usage{
call_dodo(f, ..., dodoCheck = FALSE)
}
\arguments{
\item{f}{the function to call}

\item{...}{params for f}

\item{dodoCheck}{check if a connection to DODO exists (default: FALSE).}
}
\value{
The output of the called function.
}
\description{
Call a function on the DODO graph
}
\seealso{
\link{check_dodo_connection}
}
