% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_dodo.R
\name{call_dodo_cache}
\alias{call_dodo_cache}
\title{Cached neo4j call}
\usage{
call_dodo_cache(..., tn, recache = FALSE)
}
\arguments{
\item{...}{params for \code{\link[=call_dodo]{call_dodo()}}}

\item{tn}{the name of the cached table}

\item{recache}{boolean indicating if the CQL query should be run even if
the table is already in cache}
}
\value{
The results of the \code{\link[=call_dodo]{call_dodo()}}.
}
\description{
This function calls neo4j DB the first time a query is sent and puts
the result in the cache file. The next time the same query is
called, it loads the results directly from the cache file.
}
\details{
Use only with "row" result returned by DB request.

Internal use.
}
\seealso{
\code{\link[=cache_dodo_results]{cache_dodo_results()}}, \code{\link[=load_dodo_results]{load_dodo_results()}}, \code{\link[=call_dodo]{call_dodo()}}
}
