% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB14B1.R
\name{CB14B1}
\alias{CB14B1}
\title{Caley & Barry's (2014) "Constant" model}
\usage{
CB14B1(
  records,
  alpha = 0.05,
  init.time,
  test.time = init.time + length(records) - 1,
  n.iter = 1e+05
)
}
\arguments{
\item{records}{sighting records in \code{cbin} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired threshold level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} credible interval.}

\item{init.time}{start of the observation period.}

\item{test.time}{time point to retrospectively calculate extinction
probability at. Defaults to the end of the observation period.}

\item{n.iter}{number of iterations to run (defaults to 11000).}
}
\value{
a \code{list} object with the original parameters and the p(extant),
point estimate, and credible interval included as elements. The credible
interval is a two-element numeric vector called \code{cred.int}.
}
\description{
Constant population model from Caley & Barry 2014. Estimates a posterior
probability that the species is extant at the test time, and a point
estimate and one-tailed \eqn{1 - \alpha} credibile interval on the time of
extinction.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\examples{
# Run the fox analysis from Caley & Barry 2014
CB14B1(fox, init.time = 2001, test.time = 2012)
# Run an example analysis using the Slender-billed Curlew data
CB14B1(curlew$cbin, init.time = 1817)

}
\references{
\strong{Key Reference}

Caley, P., & Barry, S. C. (2014). Quantifying extinction probabilities from
sighting records: inference and uncertainties. \emph{PLoS One}, 9(4), e95857.
\doi{10.1371/journal.pone.0095857}
}
\seealso{
\code{\link[=CB14B2]{CB14B2()}}
}
