% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-utilities.R
\name{check_df_to_import}
\alias{check_df_to_import}
\title{Check data frame to import in DODO}
\usage{
check_df_to_import(toImport, tlc, mandatory)
}
\arguments{
\item{toImport}{the object to check}

\item{tlc}{a named vector of type per column name}

\item{mandatory}{a vector of mandatory columns}
}
\description{
Feeding helper function not exported
}
\details{
Fails if conditions are not fulfilled
}
