% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_divider}
\alias{check_divider}
\title{helper function to deal with id construction}
\usage{
check_divider(ids, sep = ":", format = c("in", "out"))
}
\arguments{
\item{ids}{String. Vector of identifiers to check the format for}

\item{sep}{String. Separator used, default: ":".}

\item{format}{String. "in" for performing query in graph, "out" for converting back the
identifiers to the user}
}
\value{
vector of identifier with DB and short identifier separated by pipe or back
to colon
}
\description{
This helper function will replace the original separator by a pipe and back.
}
