% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_dodo.R
\name{check_dodo_cache}
\alias{check_dodo_cache}
\title{Check DODO cache}
\usage{
check_dodo_cache(newCon = FALSE)
}
\arguments{
\item{newCon}{if TRUE for the loading of the system information file}
}
\description{
This function checks information recorded into DODO cache
and resets it if not relevant.
}
\details{
Internal use.
}
\seealso{
\link{clear_dodo_cache}, \link{ls_dodo_cache}
}
