% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_dodo_connection}
\alias{check_dodo_connection}
\title{Check if there is a connection to a DODO database}
\usage{
check_dodo_connection(verbose = T)
}
\arguments{
\item{verbose}{if TRUE print information about the DODO connection
(default: TRUE).}
}
\value{
\itemize{
\item TRUE if the connection can be established
\item Or FALSE if the connection cannot be established
}
}
\description{
Check if there is a connection to a DODO database
}
\seealso{
\link{connect_to_dodo}
}
