% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_dodo.R
\name{clear_dodo_cache}
\alias{clear_dodo_cache}
\title{Clear the DODO cache SQLite database}
\usage{
clear_dodo_cache(queries = NULL, force = FALSE, hard = FALSE, verbose = FALSE)
}
\arguments{
\item{queries}{a character vector of the names of queries to remove.
If NULL all queries are removed.}

\item{force}{if TRUE clear the DODO cache table even if cache
file is not found}

\item{hard}{if TRUE remove everything in cache without checking file names}

\item{verbose}{display some information during the process}
}
\description{
Clear the DODO cache SQLite database
}
\seealso{
\link{ls_dodo_cache}
}
