% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_disNet.R
\name{cluster_disNet}
\alias{cluster_disNet}
\title{Cluster a disNet}
\usage{
cluster_disNet(disNet, clusterOn = c("xref"), ambiguity = 1)
}
\arguments{
\item{disNet}{the disease network to cluster and split}

\item{clusterOn}{clustering taking either or both descendants and cross-references
(values = "child", "parent", and/or "xref")}

\item{ambiguity}{level of backward ambiguity allowed
(default: 1 ==> no ambiguity allowed)}
}
\value{
A list of normalized disease networks: setDisNet object
}
\description{
Cluster a disNet object based on cross-references and/or parend/child information.
}
\details{
Cluster a disNet way generate a \emph{setDisNet} object. Based on the \emph{igraph} package to identify
connected components in a graph. These connected identifiers can identified using
cross-reference or hierarchical edges.
}
\examples{
disNet <- build_disNet(term = "psoriasis")
setDisNet <- cluster_disNet(disNet = disNet, clusterOn = "xref")

}
\seealso{
\code{\link{build_disNet}}
}
