% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-utilities.R
\name{color_database}
\alias{color_database}
\title{Database colours}
\usage{
color_database(disNet)
}
\arguments{
\item{disNet}{disNet object}
}
\description{
Function to return set colours for disease ontology resources
}
