% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_concept.R
\name{conceptSearch}
\alias{conceptSearch}
\title{Shiny server module to search concepts}
\usage{
conceptSearch(input, output, session, internal = TRUE)
}
\value{
disNet
}
\description{
Shiny module to browse diseases and phenotypes by id (db:id) or term
}
