% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_concept.R
\name{conceptSearchInput}
\alias{conceptSearchInput}
\title{Shiny ui module to search concepts}
\usage{
conceptSearchInput(id, label = "Enter search term", internal = TRUE)
}
\arguments{
\item{id}{id}

\item{label}{search term}

\item{internal}{Only provide reference ontology MONDO (TRUE) or allow choice in ontologies (FALSE)}
}
\value{
disNet
}
\description{
Shiny module to browse diseases and phenotypes by id (db:id) or term
}
