% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{connectToDODO}
\alias{connectToDODO}
\title{Connect to a neo4j DODO database}
\usage{
connectToDODO(
  url = NULL,
  username = NULL,
  password = NULL,
  connection = 1,
  remember = TRUE,
  importPath = NA
)
}
\arguments{
\item{url}{a character string. The host and the port are sufficient
(e.g: "localhost:7474")}

\item{username}{a character string (if NA ==> no auth)}

\item{password}{a character string (if NA ==> no auth)}

\item{connection}{the id of the connection already registered to use. By
default the first registered connection is used.}

\item{remember}{if TRUE the connection is registered. All the registered
connections can be listed with \link{ls_dodo_connections} and any of
them can be forgotten with \link{forget_dodo_connection}.}

\item{importPath}{the path to the import folder for loading information
in DODO (used only when feeding the database ==> default: NA)}
}
\value{
This function does not return any value. It prepares the DODO
environment to allow transparent DB calls.
}
\description{
Connect to a neo4j DODO database
}
\details{
Be carefull that you should reconnect to DODO database each time
the environment is reloaded.
}
\seealso{
\link{check_dodo_connection}, \link{ls_dodo_connections},
\link{forget_dodo_connection}
}
