% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curlew}
\alias{curlew}
\title{Slender-billed Curlew sightings (processed)}
\format{
a \code{list} with 13 elements.
}
\usage{
curlew
}
\description{
Slender-billed Curlew (\emph{Numenius tenuirostris}) sighting record from
Buchanan et al. (2025), processed into various \code{dodo} formats. The temporal
range (\code{init.time} to \code{test.time}) of these data is 1817 to 2022. Formats:
\code{ccon}, \code{cbin}, \code{cdis}, \code{ucon}, \code{ubin}, \code{udis}, \code{umcb}, \code{iucn}. An additional
"format", \code{buchanan}, contains the data in the exact \code{iucn} format as used
in Buchanan et al. (2025).
}
\references{
\strong{Key Reference}

Buchanan, G. M., Chapple, B., Berryman, A. J., Crockford, N., Jansen, J. J.
F. J., & Bond, A. L. (2025). Global extinction of Slender-billed Curlew
(\emph{Numenius tenuirostris}). \emph{Ibis}, 167(2), 357-370. \doi{10.1111/ibi.13368}
}
\seealso{
\link{curlew_raw}; \code{\link[=convert_dodo]{convert_dodo()}}; \link{curlew_effort}
}
\keyword{datasets}
