% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curlew_effort}
\alias{curlew_effort}
\title{Slender-billed Curlew effort}
\format{
a \code{numeric} vector with 206 values.
}
\usage{
curlew_effort
}
\description{
Sampling effort proxy data for the Slender-billed Curlew
(\emph{Numenius tenuirostris}). Following McCarthy et al. (1998), effort is
approximated by the number of species recorded in the species' range in each
year. Species records come from GBIF (all records of Aves with coordinates),
and range is defined as all of the countries  with 3+ confirmed (p_ci >= 0.9)
records in Buchanan et al. (2025). The temporal range (\code{init.time} to
\code{test.time}) of these data is 1817 to 2022.
}
\references{
\strong{Key Reference}

Buchanan, G. M., Chapple, B., Berryman, A. J., Crockford, N., Jansen, J. J.
F. J., & Bond, A. L. (2025). Global extinction of Slender-billed Curlew
(\emph{Numenius tenuirostris}). \emph{Ibis}, 167(2), 357-370. \doi{10.1111/ibi.13368}

\strong{Other References}

McCarthy, M. A. (1998). Identifying declining and threatened species with
museum data. Biological Conservation, 83(1), 9-17.
\doi{10.1016/S0006-3207(97)00048-7}
}
\seealso{
\link{curlew}
}
\keyword{datasets}
