% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curlew_raw}
\alias{curlew_raw}
\title{Slender-billed Curlew sightings (raw)}
\format{
a \code{data.frame} with 6 columns and 1174 rows.
}
\usage{
curlew_raw
}
\description{
Unprocessed Slender-billed Curlew (\emph{Numenius tenuirostris}) sighting record
from Buchanan et al. (2025).
}
\references{
\strong{Key Reference}

Buchanan, G. M., Chapple, B., Berryman, A. J., Crockford, N., Jansen, J. J.
F. J., & Bond, A. L. (2025). Global extinction of Slender-billed Curlew
(\emph{Numenius tenuirostris}). \emph{Ibis}, 167(2), 357-370. \doi{10.1111/ibi.13368}
}
\seealso{
\link{curlew}; \code{\link[=convert_dodo]{convert_dodo()}}
}
\keyword{datasets}
