% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curlew_surveys}
\alias{curlew_surveys}
\title{Slender-billed Curlew surveys}
\format{
a \code{data.frame} with 11 columns and 206 rows.
}
\usage{
curlew_surveys
}
\description{
Slender-billed Curlew (\emph{Numenius tenuirostris}) survey data from Buchanan et
al. (2025). The temporal range (\code{init.time} to \code{test.time}) of these data
is 1892 to 2022.
}
\references{
\strong{Key Reference}

Buchanan, G. M., Chapple, B., Berryman, A. J., Crockford, N., Jansen, J. J.
F. J., & Bond, A. L. (2025). Global extinction of Slender-billed Curlew
(\emph{Numenius tenuirostris}). \emph{Ibis}, 167(2), 357-370. \doi{10.1111/ibi.13368}
}
\keyword{datasets}
