% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{describe_concept}
\alias{describe_concept}
\title{Get concept description}
\usage{
describe_concept(ids, verbose = FALSE)
}
\arguments{
\item{ids}{a character vector of  concept identifier to search (e.g. "MONDO:0005027")}

\item{verbose}{return query content (default = FALSE)}
}
\value{
vector of disease labels
}
\description{
Describing provided concept identifiers (when available) using disease label
}
