% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{describeConcept}
\alias{describeConcept}
\title{========================================================================================

Get concept description}
\usage{
describeConcept(ids)
}
\arguments{
\item{ids}{a character vector of  concept identifier to search (e.g. "MONDO:0005027")}
}
\value{
vector of disease labels
}
\description{
Describing provided concept identifiers (when available) using disease label
}
