% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDisNet.R
\name{dim.disNet}
\alias{dim.disNet}
\title{Returns the dimension of the nodes dataframe in a disNet object}
\usage{
\method{dim}{disNet}(x, ...)
}
\description{
Returns the dimension of the nodes dataframe in a disNet object
}
