% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dodos}
\alias{dodos}
\title{Dodo sightings}
\format{
A \code{numeric} vector with 10 values.
}
\usage{
dodos
}
\description{
Dodo (\emph{Raphus cucullatus}) sighting record from Roberts & Solow (2003). The
temporal range (\code{init.time} to \code{test.time}) of these data is 1598 to 2002.
Format: \code{ccon.}
}
\references{
\strong{Key Reference}

Roberts, D. L., & Solow, A. R. (2003). Flightless birds: when did the dodo
become extinct? \emph{Nature}, 426(6964), 245. \doi{10.1038/426245a}
}
\keyword{datasets}
