% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.get_tk_headers}
\alias{.get_tk_headers}
\title{use azure token}
\usage{
.get_tk_headers(token, always = F)
}
\arguments{
\item{token}{String. Azure token}

\item{always}{Boolean.}
}
\description{
use azure token
}
