% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eskimo_curlew}
\alias{eskimo_curlew}
\title{Eskimo Curlew sightings}
\format{
a \code{data.frame} with 2 columns and 49 rows.
}
\usage{
eskimo_curlew
}
\description{
Eskimo Curlew (\emph{Numenius borealis}) sighting records from Jarić & Roberts
(2014) based on data from Roberts et al. (2010).
}
\note{
Sighting certainty values are the mean values of the reliability intervals
(i.e. 0.85, 0.7, 0.25).
}
\references{
\strong{Key Reference}

Jarić, I., & Roberts, D. L. (2014). Accounting for observation reliability
when inferring extinction based on sighting records. \emph{Biodiversity and}
\emph{Conservation}, 23(11), 2801-2815. \doi{10.1007/s10531-014-0749-8}

\strong{Other References}

Roberts, D. L., Elphick, C. S., & Reed, J. M. (2010). Identifying anomalous
reports of putatively extinct species and why it matters. \emph{Conservation}
\emph{Biology}, 24(1), 189-196.\doi{10.1111/j.1523-1739.2009.01292.x}
}
\keyword{datasets}
