% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-utilities.R
\name{explore_disNet}
\alias{explore_disNet}
\title{Explore diseases in a disease network}
\usage{
explore_disNet(disNet, show = "label", terms = NULL)
}
\arguments{
\item{disNet}{a disease network}

\item{show}{the description to be returned (label, synonym, definition)}

\item{terms}{the terms to be highlighted}
}
\value{
A \code{\link{datatable}} object.
}
\description{
Explore a disNet by a datatable highlighting a particular
term within the information provided by label, synonym or definition.
}
\examples{
exploreDisNet(extDisNet, show = c("label"), terms = "epilep")

}
