% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendDisNet.R
\name{extendDisNet}
\alias{extendDisNet}
\title{Extend a disease network}
\usage{
extendDisNet(
  disNet,
  relations = "xref",
  step = NULL,
  FA.transitivity = NULL,
  BA.transitivity = 1,
  FA.non_transitivity = NULL,
  BA.non_transitivity = NULL,
  avoidNodes = NULL,
  avoidOrigin = NULL
)
}
\arguments{
\item{disNet}{a disease network}

\item{relations}{the kind of relationships to take into account.
Possible values among combinations of: "xref", "child", "parent", and/or "pheno".
(Default value: "xref")}

\item{step}{the number of steps to take in the search (default: 10000 ==> ~exhaustive)}

\item{FA.transitivity}{forward ambiguity while using transitivity to identify cross-references (default: no filter)}

\item{BA.transitivity}{backward ambiguity while using transitivity to identify cross-references (default: 1)}

\item{FA.non_transitivity}{forward ambiguity while using transitivity to identify cross-references (default: no filter)}

\item{BA.non_transitivity}{backward ambiguity while using transitivity to identify cross-references (default: no filter)}

\item{avoidNodes}{a vector of disease ids to be avoided in the search path}

\item{avoidOrigin:}{allows to avoid traversing parent/child edges from a particular ontology (default = NULL)}

\item{verbose}{show query input (default = FALSE)}
}
\value{
A normalized disease network
}
\description{
Extending a disNet through specified relationships usingw a
specified depth. Ambiguity of cross-references can be taken into account
forward and/or backward (probably more relevant than forward).
Optionally a set of disease nodes can be avoided during the extension.
}
\details{
Once a initial disNet has been constructed, it can be extended through cross-reference or hierarchical information
through a defined number of steps (default is 10,000 steps). While extending, a vector of identifiers can be provided
through which one does not want to traverse (\emph{avoidNodes}) or resource exluded when traversing and returning
hierarchical information (\emph{avoidOrigin}). The issue of ambiguity arises when traversing cross-reference edges as
different ontology use broad or narrow definition for diseases. Two adaptations are provided
to allow controlling the transitivity between nodes.

To deal with ambiguity between ontologie cross-reference edges, the property of forward and backward ambiguity is defined.
By default, filtering is only put in place on backward ambiguity to avoid moving from a narrow concept to a broader concept.
This can result in 'hopping' through more broader defined nodes, more distantly related to the initial concept.
Filtering on forward ambiguity is not advised by default but is equivalent to moving from a more broadly defined disease
concept to several narrower disease concepts in another ontology.

It is also possible to extend to phenotype information with some limitations.
The basis of \emph{DODO} is the objective to travers the different disease ontologies efficiently and exhaustively.
A disNet can be annotated with phenotype information using the "pheno" as additional property for \emph{relations}.
However, obtaining diseases associated to a set of phenotype identifiers cannot directly be done using the extendDisNet.
This can be done using the \emph{convertConcept} function first, then followed by \emph{extendDisNet} (see vignette).
}
\examples{
disNet <- buildDisNetByTerm("epilep",fields = c("synonym","label","definition"))
extDisNet <- extendDisNet(disNet,relations = c("child","xref"), toAvoid = "MONDO:0100033")

}
\seealso{
buildDisNet
}
