% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ferret}
\alias{ferret}
\title{Black-footed Ferret sightings}
\format{
A \code{list} with 2 elements.
}
\usage{
ferret
}
\description{
Black-footed Ferret (\emph{Mustela nigripes}) sighting record from Solow (1993)
based on data provided by the US Department of the Interior, processed into
various \code{dodo} formats. The temporal range (\code{init.time} to \code{test.time}) of
these data is either 0 to 229 (\code{ccon}) or 1972 to 1992 (\code{cdis}). Formats:
\code{ccon}, \code{cdis}.
}
\note{
Times for \code{ccon} represent months post January 1972 (e.g. June 1972 is 6).
}
\references{
\strong{Key Reference}

Solow, A. R. (1993). Inferring Extinction in a Declining Population.
\emph{Journal of Mathematical Biology}, 32(1), 79-82. \doi{10.1007/Bf00160376}
}
\keyword{datasets}
