% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ferret1}
\alias{ferret1}
\title{Black-footed Ferret sightings (continuous)}
\format{
A numeric vector with 28 values.
}
\usage{
ferret1
}
\description{
Black-footed Ferret (\emph{Mustela nigripes}) sighting records from Solow
(1993) based on data provided by the US Department of the Interior.
}
\note{
Dates represent months post January 1972 (e.g. June 1972 is 6).
}
\references{
\strong{Key Reference}

Solow, A. R. (1993). Inferring Extinction in a Declining Population.
\emph{Journal of Mathematical Biology}, 32(1), 79-82. \doi{10.1007/Bf00160376}
}
\keyword{datasets}
