% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ferret2}
\alias{ferret2}
\title{Black-footed Ferret Sightings (Discrete)}
\format{
A \code{data.frame} with 2 columns and 21 rows.
}
\usage{
ferret2
}
\description{
Black-footed Ferret (\emph{Mustela nigripes}) sighting records from Solow
(1993) based on data provided by the US Department of the Interior.
}
\references{
\strong{Key Reference}

Solow, A. R. (1993). Inferring Extinction in a Declining Population.
\emph{Journal of Mathematical Biology}, 32(1), 79-82. \doi{10.1007/Bf00160376}
}
\keyword{datasets}
