% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{find_id}
\alias{find_id}
\title{Find concept by identifier}
\usage{
find_id(ids, sep = ":")
}
\arguments{
\item{ids}{a character vector of concept to search (e.g. "MONDO:0005027")}

\item{sep}{String. Separator used, default: ":".}
}
\value{
A character vector of concept identfiers in DODO
}
\description{
Find concept by identifier
}
\examples{
\dontrun{
toRet <- find_id(ids = c("MONDO:0005027", "MONDO:0100033"))
}

}
