% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{find_term}
\alias{find_term}
\title{Find concept by term}
\usage{
find_term(term, sep = ":", fields = c("label", "synonym", "definition"))
}
\arguments{
\item{term}{a character vector of terms to search (e.g. "epilep")}

\item{sep}{String. Separator used, default: ":".}

\item{fields}{the field(s) where to look for matches (label, synonym).}
}
\value{
A character vector of concept identifiers in DODO
}
\description{
Querying for concept through search terms (label, synonym)
}
