% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findConcepts.R
\name{findID}
\alias{findID}
\title{Find concept by identifier}
\usage{
findID(ids)
}
\arguments{
\item{ids}{a character vector of concept to search formatted as "DB:id" (e.g. "MONDO:0005027")}
}
\value{
A character vector of concept identfiers in DODO
}
\description{
Find concept by identifier
}
\examples{
findID(id = c("MONDO:0005027","MONDO:0100033"))

}
