% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findConcepts.R
\name{findTerm}
\alias{findTerm}
\title{Find concept by term}
\usage{
findTerm(term, fields = c("label", "synonym", "definition"))
}
\arguments{
\item{term}{a character vector of terms to search (e.g. "epilep")}

\item{fields}{the field(s) where to look for matches (label, synonym, definition).}
}
\value{
A character vector of concept identfiers in DODO
}
\description{
Querying for concept through searchterms (label, synonym, definition)
}
