% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focus_disNet.R
\name{focus_disNet}
\alias{focus_disNet}
\title{Focus a disNet on identifiers of interest and neighbors}
\usage{
focus_disNet(disNet, diseaseID, relationship = c("xref", "child"), steps = 1)
}
\arguments{
\item{disNet}{a disease network}

\item{diseaseID}{a character vector of disease ids of interest (format "MONDO:0005027")}

\item{relationship}{a character vector of the relationships to take into account. Can take values "xref","child","parent.
Default is "xref" and "child".}

\item{steps}{the maximum number of steps between diseases of interest
and their neighbors (default: 2)}
}
\value{
A normalized disease network
}
\description{
Focussing on particular identifiers of interest while also extending the disNet using the identifiers
of interest as a starting point. Hierarchical or cross-reference edges are traversed that exist within the original
disNet. This doesn't require a connection to a database instance and will only consider the edges available in the
provided disNet object.
}
\examples{

df <- as.data.frame(id = c("ORPHA:86814","MONDO:0005099"),
                    database = c("ORPHA","MONDO"))
buildDisNetByID(df = df)
extDisNet <- extendDisNet(disNet,relations = c("child","xref"), toAvoid = "MONDO:0100033")
focusDisNet <- focusNet(disNet = extDisNet,diseaseID = "MONDO:0015955",steps = 2)

}
