% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{forget_dodo_connection}
\alias{forget_dodo_connection}
\title{Forget a DODO connection}
\usage{
forget_dodo_connection(connection)
}
\arguments{
\item{connection}{the id of the connection to forget.}
}
\description{
Forget a DODO connection
}
\seealso{
\link{ls_dodo_connections},
\link{connect_to_dodo}, \link{check_dodo_connection}
}
