% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_disNet.R
\name{format.disNet}
\alias{format.disNet}
\title{format disNet}
\usage{
\method{format}{disNet}(x, ...)
}
\arguments{
\item{x}{disNet object}

\item{...}{additional parameters}
}
\description{
format disNet
}
