% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_disNet.R
\name{format.setDisNet}
\alias{format.setDisNet}
\title{format disNet}
\usage{
\method{format}{setDisNet}(x, ...)
}
\arguments{
\item{x}{disNet object}

\item{...}{additional parameters}
}
\description{
format disNet
}
