% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gerygone}
\alias{gerygone}
\title{Lord Howe Gerygone sightings}
\format{
a \code{numeric} vector with 236 values.
}
\usage{
gerygone
}
\description{
Lord Howe Gerygone (\emph{Gerygone insularis}) sighting record based on data from
the Atlas of Living Australia. The temporal range (\code{init.time} to
\code{test.time}) of these data is 1788 to 2023. Format: \code{cdis}.
}
\references{
\strong{Key Reference}

Atlas of Living Australia occurrence download, accessed on 25 August 2025.
\doi{10.26197/ala.dad3870b-1b79-4600-ad12-af429feb1ad2}

\strong{Other References}

McCarthy, M. A. (1998). Identifying declining and threatened species with
museum data. \emph{Biological Conservation}, 83(1), 9-17.
\doi{10.1016/S0006-3207(97)00048-7}
}
\seealso{
\link{gerygone_effort}
}
\keyword{datasets}
