% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{get_concept_url}
\alias{get_concept_url}
\title{Get URL of database}
\usage{
get_concept_url(ids, sep = ":", exact = TRUE)
}
\arguments{
\item{ids}{concept identifiers as vector (e.g. "MONDO:0005027")}

\item{sep}{String. Separator used, default: ":".}

\item{exact}{returns url of the original database (e.g. MONDO url for "MONDO:0005027"). Default = TRUE}
}
\value{
Returns a database with dbid, db, id, database, url
}
\description{
Returns a vector of urls of the provided concept identifiers
}
\examples{
\dontrun{
get_concept_url(ids = "MONDO:0005027")
}

}
