% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{get_ontology}
\alias{get_ontology}
\title{Return entire disease ontology}
\usage{
get_ontology(database, sep = ":")
}
\arguments{
\item{database}{disease ontology to return.}

\item{sep}{String. Separator used, default: ":".}
}
\description{
Return all nodes for a particular disease ontology
}
\seealso{
\code{\link{list_database}}
}
