% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{get_version}
\alias{get_version}
\title{Returns current version}
\usage{
get_version()
}
\value{
Returns current version of DODO instance
}
\description{
Returns the version of the current DODO database
}
