% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntology.R
\name{getOntology}
\alias{getOntology}
\title{Embryonic function to get full ontology}
\usage{
getOntology(database, concept = c("Disease", "Phenotype"))
}
\description{
Embryonic function to get full ontology
}
