% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getSchema}
\alias{getSchema}
\title{=======================================================================================

Show the data model of DODO}
\usage{
getSchema()
}
\description{
Show the shema of the DODO data model.
}
