% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getVersion}
\alias{getVersion}
\title{========================================================================================

Returns current version}
\usage{
getVersion()
}
\value{
Returns current version of DODO instance
}
\description{
Returns the version of the current DODO database
}
\examples{
getVersion()

}
