% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grebe_surveys}
\alias{grebe_surveys}
\title{Alaotra Grebe surveys}
\format{
a \code{data.frame} with 11 columns and 89 rows.
}
\usage{
grebe_surveys
}
\description{
Alaotra Grebe (\emph{Tachybaptus rufolavatus}) survey data from Thompson et
al. (2017). The temporal range (\code{init.time} to \code{test.time}) of these data
is 1929 to 2017.
}
\references{
\strong{Key Reference}

Thompson, C. J., Koshkina, V., Burgman, M. A., Butchart, S. H. M., & Stone,
L. (2017). Inferring extinctions II: A practical, iterative model based on
records and surveys. \emph{Biological Conservation}, 214, 328-335.
\doi{10.1016/j.biocon.2017.07.029}
}
\keyword{datasets}
