% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ground_dove}
\alias{ground_dove}
\title{Purple-winged Ground Dove sightings}
\format{
A \code{data.frame} with 2 columns and 61 rows.
}
\usage{
ground_dove
}
\description{
Purple-winged Ground Dove (\emph{Paraclaravis geoffroyi}) sighting records from
Lees et al. (2021).
}
\note{
Adapted from Supplementary Table 1 in Lees et al. (2021). Almost all specimen
records are considered certain, while sight and literature-based records are
all considered uncertain. The controversial specimen from 1991 is classified
as uncertain. Records have been deduplicated to ensure they are all
independent from one another. Records without a date are not included.
}
\references{
\strong{Key Reference}

Lees, A. C., Devenish, C., Areta, J. I., de Araújo, C. B., Keller, C.,
Phalan, B., & Silveira, L. F. (2021). Assessing the Extinction Probability
of the Purple-winged Ground Dove, an Enigmatic Bamboo Specialist.
\emph{Frontiers in Ecology and Evolution}, 9. \doi{10.3389/fevo.2021.624959}
}
\keyword{datasets}
