% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-utilities.R
\name{harmonize_Cortellis}
\alias{harmonize_Cortellis}
\title{========================================================================================

Harmonize Cortellis information}
\usage{
harmonize_Cortellis(
  CortellisONT_siConditions,
  CortellisONT_siConditionParents,
  CortellisONT_siConditionXref,
  CortellisONT_ciIndications,
  CortellisONT_ciIndicationParents,
  CortellisONT_ciIndicationXref,
  checkMissing = NULL
)
}
\arguments{
\item{CortellisONT_siConditions}{dataframe from CortellisONT siCondition with id (conditionId) and
name (disease label)}

\item{CortellisONT_siConditionParents}{dataframe from CortellisONT siCondition with id (conditionId),
depth (depth in the hierarchy), leaf, treeCode, parent (parent identifier)}

\item{CortellisONT_siConditionXref}{dataframe from CortellisONT siCondition with conditionId, source,
xref, dbRef, direct}

\item{CortellisONT_ciIndications}{dataframe from CortellisONT ciIndication with id (indicationId) and
name (disease label)}

\item{CortellisONT_ciIndicationParents}{dataframe from CortellisONT ciIndication with id (indicationId),
depth (depth in the hierarchy), leaf, treeCode, parent (parent identifier)}

\item{CortellisONT_ciIndicationXref}{dataframe from CortellisONT ciIndication with indicationId, source,
xref, dbRef, direct}

\item{tkcon}{Connection to TKCat instance to check for missing identifiers}
}
\value{
list with harmonized dataframes
}
\description{
Harmonizing the Cortellis input information for DODO
}
