% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-utilities.R
\name{harmonize_HPO}
\alias{harmonize_HPO}
\title{Harmonize HPO}
\usage{
harmonize_HPO(
  HPO_diseaseHP,
  HPO_synonyms,
  HPO_hp,
  HPO_altId,
  HPO_diseases,
  HPO_parents,
  Monarch_hp
)
}
\arguments{
\item{HPO_diseaseHP}{Dataframe connecting disease identifiers to HPO identifiers
(columns: db - disease identifier database, id - short disease identifier, hp - short HP identifier)}

\item{HPO_synonyms}{Dataframe with id (HP short identifier), synonym, and type (of synonym)}

\item{HPO_hp}{Dataframe with columns id (HP short identifier), name (label information),
description (description of phenotype), level (in hierarchy)}

\item{HPO_altId}{Dataframe with columns id (HP short identifier) and alt (alternative HP short identifier)}

\item{HPO_diseases}{Dataframe with columns db (disease identifier database), id (disease short identifier),
label (disease label)}

\item{HPO_parents}{Dataframe with columns id and parent}

\item{Monarch_hp}{Dataframe from Monarch Initiative linking MONDO ids to HPO ids (columns DB, id, hp)}
}
\description{
Harmoning the parsed files from HPO before merge
}
