% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{importinDODO}
\alias{importinDODO}
\title{Feeding DODO: Imports a data.frame in the DODO graph database}
\usage{
importinDODO(cql, toImport, ...)
}
\arguments{
\item{cql}{the CQL query to be applied on each row of toImport}

\item{toImport}{the data.frame to be imported as "row".
Use "row.FIELD" in the cql query to refer to one FIELD of the toImport
data.frame}

\item{...}{additional parameters to the \code{\link[neo2R:import_from_df]{neo2R::import_from_df()}} function.}
}
\value{
the results of the query
}
\description{
Not exported to avoid unintended modifications of the DB.
}
\seealso{
\code{\link[=call_dodo]{call_dodo()}}
}
