% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_concept.R
\name{indirect_cql}
\alias{indirect_cql}
\title{helper function for indirect cql query}
\usage{
indirect_cql(
  edge_types = edge_types,
  to_db = to_db,
  relationship = relationship,
  network = F
)
}
\arguments{
\item{edge_types}{List. Edge types to query. See \link{convert_concept}}

\item{to_db}{String. Database to filter on.}

\item{relationship}{String. Relationship to filter for.}

\item{network}{Boolean. Network or data.table to be returned as output.}
}
\description{
will return the cql query to use the apoc.path.expandConfig
based on a data.table or network to be returned
}
