% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_disNet.R
\name{is.setDisNet}
\alias{is.setDisNet}
\title{Check validity of setDisNet object}
\usage{
is.setDisNet(x, ...)
}
\arguments{
\item{x}{disNet object}

\item{...}{additional parameters}
}
\description{
Check validity of setDisNet object
}
