% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JR14F1.R
\name{JR14F1}
\alias{JR14F1}
\title{Jarić & Roberts' (2014) "Strauss & Sadler 1989" Model}
\usage{
JR14F1(records, alpha = 0.05, init.time = min(records$time))
}
\arguments{
\item{records}{\code{data.frame} with two columns: \code{time} and \code{certainty}. The
\code{time} column contains the date of all sightings, which each have an
associated \code{certainty}, expressed as a probability in the interval [0, 1]
(values close to 1 imply a high likelihood that the observation is correct).}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}

\item{init.time}{start of the observation period. Defaults to the time of
the first sighting.}
}
\value{
a \code{list} object with the original parameters and the confidence
interval included as elements. The confidence interval is a two-element
numeric vector called \code{conf.int}.
}
\description{
Modification of the classical confidence interval presented by Strauss &
Sadler 1989 from Jarić & Roberts 2014. Estimates a one-tailed
\eqn{1 - \alpha} confidence interval on the time of
extinction. Sighting uncertainty is incorporated.
}
\note{
Sampling effort is assumed to be constant.
}
\references{
\strong{Key Reference}

Jarić, I., & Roberts, D. L. (2014). Accounting for observation reliability
when inferring extinction based on sighting records. \emph{Biodiversity and}
\emph{Conservation}, 23(11), 2801-2815. \doi{10.1007/s10531-014-0749-8}

\strong{Other References}

Strauss, D., & Sadler, P. M. (1989). Classical Confidence Intervals and
Bayesian Probability Estimates for Ends of Local Taxon Ranges. Mathematical
Geology, 21(4), 411-421. \doi{10.1007/Bf00897326}
}
\seealso{
\code{\link[=JR14F2]{JR14F2()}}, \code{\link[=JR14F3]{JR14F3()}}, \code{\link[=JR14F4]{JR14F4()}}
}
