% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JR14F4.R
\name{JR14F4}
\alias{JR14F4}
\title{Jarić & Roberts' (2014) "McInerny et al. 2006" model}
\usage{
JR14F4(records, alpha = 0.05, init.time = min(records$time))
}
\arguments{
\item{records}{\code{data.frame} with two columns: \code{time} and \code{certainty}. The
\code{time} column contains the date of all sightings, which each have an
associated \code{certainty}, expressed as a probability in the interval [0, 1]
(values close to 1 imply a high likelihood that the observation is correct).}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}

\item{init.time}{start of the observation period. Defaults to the time of
the first sighting.}
}
\value{
a \code{list} object with the original parameters and the confidence
interval included as elements. The confidence interval is a two-element
numeric vector called \code{conf.int}.
}
\description{
Modification of the confidence interval presented by McInerny et al. 2006
from Jarić & Roberts 2014. Estimates a one-tailed \eqn{1 - \alpha}
confidence interval on the time of extinction. Sighting uncertainty is
incorporated.
}
\note{
Sampling effort is assumed to be constant.
}
\references{
\strong{Key Reference}

Jarić, I., & Roberts, D. L. (2014). Accounting for observation reliability
when inferring extinction based on sighting records. \emph{Biodiversity and}
\emph{Conservation}, 23(11), 2801-2815. \doi{10.1007/s10531-014-0749-8}

\strong{Other References}

McInerny, G. J., Roberts, D. L., Davy, A. J., & Cribb, P. J. (2006).
Significance of sighting rate in inferring extinction and threat.
\emph{Conservation Biology}, 20(2), 562-567.
\doi{10.1111/j.1523-1739.2006.00377.x}
}
\seealso{
\code{\link[=JR14F1]{JR14F1()}}, \code{\link[=JR14F2]{JR14F2()}}, \code{\link[=JR14F3]{JR14F3()}}
}
