% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KO21B1.R
\name{KO21B1}
\alias{KO21B1}
\title{Kodikara et al.'s (2021) "Change-point" model}
\usage{
KO21B1(
  records,
  alpha = 0.05,
  init.time,
  test.time = init.time + nrow(records) - 1
)
}
\arguments{
\item{records}{sighting records in \code{ubin} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired threshold level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} credible interval.}

\item{init.time}{start of the observation period.}

\item{test.time}{time point to retrospectively calculate extinction
probability at. Defaults to the end of the observation period.}
}
\value{
a \code{list} object with the original parameters and the p(extant),
point estimate, and credible interval included as elements. The credible
interval is a two-element numeric vector called \code{cred.int}.
}
\description{
The model from Kodikara et al. 2021. Estimates a posterior probability that
the species is extant at the test time, and a point estimate and one-sided
\eqn{1 - \alpha} credible interval on the time of extinction.
}
\note{
Sampling effort is assumed to be constant.
}
\examples{
\dontrun{
# Run the Ivory-billed Woodpecker analysis from Kodikara et al. 2021
KO21B1(woodpecker$ubin, init.time = 1897, test.time = 2010)
# Run an example analysis using the Slender-billed Curlew data
KO21B1(curlew$ubin, init.time = 1817, test.time = 2022)
}

}
\references{
\strong{Key Reference}

Kodikara, S., Demirhan, H., Wang, Y., Solow, A., & Stone, L. (2020).
Inferring extinction year using a Bayesian approach. \emph{Methods in Ecology}
\emph{and Evolution}, 11(8), 964-973. \doi{10.1111/2041-210x.13408}
}
