% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laughing_owl}
\alias{laughing_owl}
\title{Laughing Owl sightings}
\format{
a \code{list} with 12 elements.
}
\usage{
laughing_owl
}
\description{
Laughing Owl (\emph{Ninox albifacies}) sighting record from Butchart et al.
(2018), processed into various \code{dodo} formats. The temporal range
(\code{init.time} to \code{test.time}) of these data is 1844 to 2018 Formats:
\code{ccon}, \code{cbin}, \code{cdis}, \code{ucon}, \code{ubin}, \code{udis}, \code{umcb}, \code{iucn}.
}
\references{
\strong{Key Reference}

Butchart, S. H., Lowe, S., Martin, R. W., Symes, A., Westrip, J. R., &
Wheatley, H. (2018). Which bird species have gone extinct? A novel
quantitative classification approach. \emph{Biological Conservation}, 227, 9-18.
\doi{10.1016/j.biocon.2018.08.014}
}
\seealso{
\link{laughing_owl_effort}
}
\keyword{datasets}
