% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LE14B1.R
\name{LE14B1}
\alias{LE14B1}
\title{Lee's (2014) "Simple" model}
\usage{
LE14B1(records, surveys, threshold = 0.9, prior = c(0, 1), n.iter = 10000)
}
\arguments{
\item{records}{sighting records in \code{iucn} format (see
\code{\link{convert_dodo}} for details).}

\item{surveys}{survey effort data (passive survey effort is ignored). A
\code{data.frame} with 11 columns: \code{time}, \code{survey}, \code{epsilon}, \code{epsilon_lower},
\code{epsilon_upper}, \code{p_i}, \code{p_i_lower}, \code{p_i_upper}, \code{p_r}, \code{p_r_lower}, and
\code{p_r_upper}. The number of rows must match \code{records}.}

\item{threshold}{cutoff certainty value for which sightings to consider
certain. Defaults to \code{0.9}.}

\item{prior}{a two-element \code{numeric} vector with the lower and upper bounds
of the (uniform) prior for extant (P(X_T)). Defaults to \code{c(0, 1)}.}

\item{n.iter}{number of iterations to run (defaults to 10,000).}
}
\value{
a \code{list} object with the original parameters and p(extant) included
as elements.
}
\description{
The simple numerical tool presented in Lee 2014, including both uncertain
sightings and surveys. Estimates a posterior probability that the species is
extant at the end of the observation period.
}
\note{
Passive survey effort (rows where \code{survey == FALSE}) are ignored, unlike in
\code{\link[=TH17I1]{TH17I1()}}.
}
\examples{
# Run the Pohnpei Mountain Starling analysis from Lee 2014.
LE14B1(
  records = starling, surveys = starling_surveys, threshold = 0.9,
  prior = c(0.24, 0.48), n.iter = 1e4
)
# Run an example analysis using the Alaotra Grebe data
\dontrun{
LE14B1(
  records = grebe, surveys = grebe_surveys, threshold = 0.9,
  prior = c(0, 1), n.iter = 1e4
)
}

}
\references{
\strong{Key Reference}

Lee, T. E. (2014). A simple numerical tool to infer whether a species is
extinct. \emph{Methods in Ecology and Evolution}, 5(8), 791-796.
\doi{10.1111/2041-210x.12227}
}
