% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LE14B2.R
\name{LE14B2}
\alias{LE14B2}
\title{Lee et al.'s (2014) "Variable reliability" model}
\usage{
LE14B2(
  records,
  alpha = 0.05,
  init.time,
  n.chains = 4,
  n.iter = 11000,
  n.burnin = 1000,
  n.thin = 10
)
}
\arguments{
\item{records}{sighting records in \code{ubin} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired threshold level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} credible interval.}

\item{init.time}{start of the observation period.}

\item{n.chains}{number of MCMC chains to run. Defaults to 4.}

\item{n.iter}{number of iterations in each chain. Defaults to 11000.}

\item{n.burnin}{number of iterations to discard as burn-in. Defaults to 1000.}

\item{n.thin}{thinning rate. Defaults to 10.}
}
\value{
a \code{list} object with the original parameters and the p(extant),
point estimate, and credible interval included as elements. The credible
interval is a two-element numeric vector called \code{cred.int}.
}
\description{
The model from Lee et al. 2014, incorporating uncertain sightings. Estimates
a posterior probability that the species is extant at the end of the
observation period, and a point estimate and one-tailed \eqn{1 - \alpha}
credible interval on the time of extinction, conditional on extinction
occurring before the end of the observation period.
}
\note{
Sampling effort is assumed to be constant.
}
\examples{
# Run the example analysis from Lee et al. 2014
LE14B2(lee_s1, init.time = 1)
\dontrun{
# Run an example analysis using the Slender-billed Curlew data
LE14B2(curlew$ubin, init.time = 1817)
}

}
\references{
\strong{Key Reference}

Lee, T. E., McCarthy, M. A., Wintle, B. A., Bode, M., Roberts, D. L., &
Burgman, M. A. (2014). Inferring extinctions from sighting records of
variable reliability. \emph{Journal of Applied Ecology}, 51(1), 251-258.
\doi{10.1111/1365-2664.12144}
}
