% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lee_s1}
\alias{lee_s1}
\title{Example sightings from Lee et al.}
\format{
a \code{data.frame} with 2 columns and 9 rows.
}
\usage{
lee_s1
}
\description{
Example binary sighting record from Appendix S1 in Lee et al. (2014). The
temporal range (\code{init.time} to \code{test.time}) of these data is 1 to 9.
Format: \code{ubin}.
}
\references{
\strong{Key Reference}

Lee, T. E., McCarthy, M. A., Wintle, B. A., Bode, M., Roberts, D. L., &
Burgman, M. A. (2014). Inferring extinctions from sighting records of
variable reliability. \emph{Journal of Applied Ecology}, 51(1), 251-258.
\doi{10.1111/1365-2664.12144}
}
\keyword{datasets}
