% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_database}
\alias{list_database}
\title{List of databases in DODO}
\usage{
list_database()
}
\value{
Returns a list of all database in DODO
}
\description{
lists all databases present in the database
}
\examples{
list_database()

}
