% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_db}
\alias{list_db}
\title{List of databases in DODO}
\usage{
list_db()
}
\value{
Returns a list of all database in DODO
}
\description{
lists all databases present in the database
}
\examples{
listDB()

}
