% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_node_type}
\alias{list_node_type}
\title{List of nodes types in DODO}
\usage{
list_node_type()
}
\value{
Returns a list of all database in DODO
}
\description{
Lists the different type of node present in the database
}
\examples{
list_node_type()

}
