% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{listConnections}
\alias{listConnections}
\title{List all registered DODO connection}
\usage{
listConnections()
}
\description{
List all registered DODO connection
}
\seealso{
\link{connect_to_dodo},
\link{forget_dodo_connection}, \link{check_dodo_connection}
}
