% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{listType}
\alias{listType}
\title{========================================================================================

List of nodes types in DODO}
\usage{
listType()
}
\value{
Returns a list of all database in DODO
}
\description{
Lists the different type of node present in the database
}
\examples{
listType()

}
