% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feed-load_alternative_identifiers.R
\name{load_alternative_identifiers}
\alias{load_alternative_identifiers}
\title{Feeding DODO: load concept alternative identifiers in DODO DB}
\usage{
load_alternative_identifiers(toImport, concept)
}
\arguments{
\item{toImport}{a data.frame with the following columns:
\itemize{
\item database: \emph{character} name of the database (mandatory)
\item shortID: \emph{character} short concept ID (mandatory)
\item altdb: \emph{character} name of the database for
alternative identifiers (mandatory)
\item altid: \emph{character} alternative short concept identifiers (mandatory)
}}

\item{concept}{either "Disease" or "Phenotype"}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
