% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-utilities.R
\name{load_ClinVar}
\alias{load_ClinVar}
\title{Load Clinvar disease IDs}
\usage{
load_ClinVar(DODO_entryId, path_name, traitCref, traitName)
}
\arguments{
\item{DODO_entryId}{the object DODO_entryId to filter out existing identifiers}

\item{path_name}{path to ClinVar parsed files}

\item{traitCref}{name of the file containing the dataframe with association ClinVar identifiers to external disease identifiers
(columns t.id - ClinVar identifier, db - disease database, id - disease identifier, and type)}

\item{traitName}{name of the file containing dataframe with columns t.id (ClinVar identifier) and label}
}
\value{
list with harmonized files
}
\description{
Integrate disease identifiers from the ClinVar resource
}
