% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-utilities.R
\name{load_concept_definitions}
\alias{load_concept_definitions}
\title{Feeding DODO: load concept defintions in DODO DB}
\usage{
load_concept_definitions(toImport, concept)
}
\arguments{
\item{toImport}{a data.frame with the following columns:
\itemize{
\item database: \emph{character} name of the concept reference database (mandatory)
\item origin: \emph{character} name of the database from which the concept was
taken (optional). If not provided or NA, the database field is used.
\item shortID: \emph{character} short concept ID (mandatory)
\item label: \emph{character} concept label (optional)
\item definition: \emph{character} concept definition (optional)
\item level: \emph{integer} level in database ontology (optional)
}}

\item{concept}{either "Disease" or "Phenotype"}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
