% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-utilities.R
\name{load_concept_synonyms}
\alias{load_concept_synonyms}
\title{Feeding DODO: load concept synonyms in DODO DB}
\usage{
load_concept_synonyms(toImport, concept)
}
\arguments{
\item{toImport}{a data.frame with the following columns:
\itemize{
\item database: \emph{character} name of the database (mandatory)
\item shortID: \emph{character} short concept ID (mandatory)
\item value: \emph{character} concept synonym (mandatory)
}}

\item{concept}{either "Disease" or "Phenotype"}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
