% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feed-load_cross_references.R
\name{load_cross_references}
\alias{load_cross_references}
\title{Feeding DODO: load concept cross references in DODO DB}
\usage{
load_cross_references(toImport, xrefDB, concept)
}
\arguments{
\item{toImport}{the data frame of cross references with the following column
\itemize{
\item DB1: \emph{character} name of the database 1 (mandatory)
\item id1: \emph{character} short concept ID 1 (mandatory)
\item DB2: \emph{character} name of the database 2 (mandatory)
\item id2: \emph{character} short concept ID 2 (mandatory)
}}

\item{xrefDB}{a data frame with 2 columns ("DB1", "DB2") indicating couples
of DB considered close enough to define "xref". Identifiers from other
couples of DB will be considered as being "related"}

\item{concept}{either "Disease" or "Phenotype"}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
