% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feed-load_db_definition.R
\name{load_db_definition}
\alias{load_db_definition}
\title{Feeding DODO: Register database in DODO DB}
\usage{
load_db_definition(toImport)
}
\arguments{
\item{toImport}{a data.frame with the following columns:
\itemize{
\item name: \emph{character} name of the database
\item idURL: \emph{character} URL template for database concepts
}}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
