% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feed-load_db_definitions.R
\name{load_db_definitions}
\alias{load_db_definitions}
\title{Feeding DODO: Register databases with URL template in DODO DB}
\usage{
load_db_definitions(toImport)
}
\arguments{
\item{toImport}{a data.frame with the following columns:
\itemize{
\item name: \emph{character} name of the database (mandatory)
\item idURL: \emph{character} URL template for database concepts (optional)
}}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
