% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-utilities.R
\name{load_db_names}
\alias{load_db_names}
\title{Feeding DODO: Register databases in DODO DB}
\usage{
load_db_names(toImport)
}
\arguments{
\item{toImport}{a data.frame with the following columns:
\itemize{
\item name: \emph{character} name of the database (mandatory)
}}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
