% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-utilities.R
\name{load_dodo_model}
\alias{load_dodo_model}
\title{Feeding DODO: Load DODO data model in neo4j}
\usage{
load_dodo_model()
}
\description{
Not exported to avoid unintended modifications of the DB.
}
