% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_dodo.R
\name{load_dodo_results}
\alias{load_dodo_results}
\title{Get a DODO query result from cache}
\usage{
load_dodo_results(name)
}
\arguments{
\item{name}{the name of the query}
}
\description{
Internal use
}
\seealso{
\code{\link[=call_dodo_cache]{call_dodo_cache()}}, \code{\link[=cache_dodo_results]{cache_dodo_results()}}
}
