% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-utilities.R
\name{load_has_phenotypes}
\alias{load_has_phenotypes}
\title{Feeding DODO: load diseases-phenotypes relationships in DODO DB}
\usage{
load_has_phenotypes(toImport)
}
\arguments{
\item{toImport}{a data.frame with the following columns:
\itemize{
\item diseaseDB: \emph{character} name of the disease database (mandatory)
\item diseaseID: \emph{character} short disease ID (mandatory)
\item phenoDB: \emph{character} name of the phenotype database (mandatory)
\item phenoID: \emph{character} short phenotype ID (mandatory)
}}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
