% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-utilities.R
\name{load_parent_identifiers}
\alias{load_parent_identifiers}
\title{Feeding DODO: load concept parent identifiers in DODO DB}
\usage{
load_parent_identifiers(toImport, concept, origin)
}
\arguments{
\item{toImport}{a data.frame with the following columns:
\itemize{
\item database: \emph{character} name of the database (mandatory)
\item shortID: \emph{character} short concept ID (mandatory)
\item parentdb: \emph{character} name of the database for
parent identifiers (mandatory)
\item parentid: \emph{character} parent short concept identifiers (mandatory)
}}

\item{concept}{either "Disease" or "Phenotype"}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
