% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_dodo.R
\name{ls_dodo_cache}
\alias{ls_dodo_cache}
\title{List all the DODO queries in cache and the total size of the cache}
\usage{
ls_dodo_cache(verbose = TRUE)
}
\arguments{
\item{verbose}{if TRUE (default) prints a message displaying the total
size of the cache}
}
\value{
A data.frame giving for each query (row names) its size in Bytes
(column "size") and in human readable format (column "hr"). The
attribute "Total" corresponds to the sum of all the file size.
}
\description{
List all the DODO queries in cache and the total size of the cache
}
\seealso{
\link{clear_dodo_cache}
}
