% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA94F1.R
\name{MA94F1}
\alias{MA94F1}
\title{Marshall's (1994) "Distribution-free" model}
\usage{
MA94F1(records, alpha = 0.05, gamma = 0.1)
}
\arguments{
\item{records}{sighting records in \code{ccon} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}

\item{gamma}{desired confidence probability of the bounds of the confidence
interval.}
}
\value{
a \code{list} object with the original parameters and the p-value, point
estimate, and confidence interval included as elements. The confidence
interval is a three-element numeric vector called \code{conf.int}; the first
element is the lower bound of the confidence interval (the time of the last
record), the second element is the lower bound of the upper bound of the
confidence interval, and the third element is the upper bound of the upper
bound of the confidence interval.
}
\description{
Equations 1 and 2 from Marshall 1994. Estimates a one-tailed
\eqn{1 - \alpha} confidence interval on time of extinction; the upper bound
of this confidence interval is in itself a \eqn{1 - 2\gamma} confidence
interval with a lower and upper bound.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant. The upper bound of the upper confidence interval is likely to
be NA, unless there are sufficient sightings or \eqn{\alpha} and \eqn{\gamma}
are sufficiently permissive.
}
\examples{
# Run the *Metrarabdotos* n. sp. 5 analysis from Marshall 1994
MA94F1(metrarabdotos, alpha = 0.5, gamma = 0.025)
# Run an example analysis using the Slender-billed Curlew data
MA94F1(curlew$ccon, alpha = 0.05, gamma = 0.05)

}
\references{
\strong{Key Reference}

Marshall, C. R. (1994). Confidence intervals on stratigraphic ranges:
partial relaxation of the assumption of randomly distributed fossil
horizons. \emph{Paleobiology}, 20(4), 459-469. \doi{10.1017/S0094837300012938}
}
