% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA97F1.R
\name{MA97F1}
\alias{MA97F1}
\title{Marshall's (1997) "Non-random" model}
\usage{
MA97F1(records, effort, alpha = 0.05, init.time)
}
\arguments{
\item{records}{sighting records in \code{cdis} format (see
\code{\link{convert_dodo}} for details).}

\item{effort}{a vector of effort data, of the same length as \code{records}.}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}

\item{init.time}{start of the observation period.}
}
\value{
a \code{list} object with the original parameters and the confidence
interval included as elements. The confidence interval is a two-element
numeric vector called \code{conf.int}.
}
\description{
Equations 3 to 5 from Marshall 1997. Estimates a one-sided \eqn{1 - \alpha}
confidence interval on time of extinction.
}
\note{
All sighting records are assumed to be certain. Uses the discrete form as
presented in Rivadeneira et al. (2009).
}
\examples{
# Run an example analysis using the Lord Howe Gerygone data
MA97F1(records = gerygone, effort = gerygone_effort, init.time = 1788)
# Run an example analysis using the Slender-billed Curlew data
\dontrun{
MA97F1(records = curlew$cdis, effort = curlew_effort, init.time = 1817)
}
}
\references{
\strong{Key Reference}

Marshall, C. R. (1997). Confidence intervals on stratigraphic ranges with
nonrandom distributions of fossil horizons. \emph{Paleobiology}, 23(2), 165-173.
\doi{10.1017/S0094837300016766}

\strong{Other References}

Rivadeneira, M. M., Hunt, G., & Roy, K. (2009). The use of sighting records
to infer species extinctions: an evaluation of different methods. \emph{Ecology},
90(5), 1291-1300. \doi{10.1890/08-0316.1}
}
