% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammoth}
\alias{mammoth}
\title{Alaskan Mammoth fossil record}
\format{
A numeric vector with 25 values.
}
\usage{
mammoth
}
\description{
Woolly Mammoth (\emph{Mammuthus primigenius}) dated fossil records from Solow et
al. (2006) based on data presented in Guthrie (2004).
}
\note{
Includes the 25 most recent fossil records, measured in years before present
(note that values are all <0). Radiometric dating error estimates are not
included.
}
\references{
\strong{Key Reference}

Solow, A. R., Roberts, D. L., & Robbirt, K. M. (2006). On the Pleistocene
extinctions of Alaskan mammoths and horses. \emph{Proceedings of the National}
\emph{Academy of Sciences of the United States of America}, 103(19), 7351-7353.
\doi{10.1073/pnas.0509480103}

\strong{Other References}

Guthrie, R. D. (2004). Radiocarbon evidence of mid-Holocene mammoths
stranded on an Alaskan Bering Sea island. \emph{Nature}, 429(6993), 746-749.
\doi{10.1038/nature02612}

Bradshaw, C. J. A., Cooper, A., Turney, C. S. M., & Brook, B. W. (2012).
Robust estimates of extinction time in the geological record. \emph{Quaternary}
\emph{Science Reviews}, 33, 14-19. \doi{10.1016/j.quascirev.2011.11.021}
}
\keyword{datasets}
