% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC98F1.R
\name{MC98F1}
\alias{MC98F1}
\title{McCarthy's (1998) "Partial Solow" model}
\usage{
MC98F1(records, effort)
}
\arguments{
\item{records}{sighting records in \code{cdis} format (see
\code{\link{convert_dodo}} for details).}

\item{effort}{a vector of effort data, of the same length as \code{records}.}
}
\value{
a \code{list} object with the original parameters and the p-value
included as elements.
}
\description{
Equation 2 from McCarthy 1998. Estimates a p-value for testing
competing hypotheses of extinction/non-extinction.
}
\note{
All sighting records are assumed to be certain.
}
\examples{
# Run an example analysis using the Lord Howe Gerygone data
MC98F1(records = gerygone, effort = gerygone_effort)
# Run an example analysis using the Slender-billed Curlew data
MC98F1(records = curlew$cdis, effort = curlew_effort)

}
\references{
\strong{Key Reference}

McCarthy, M. A. (1998). Identifying declining and threatened species with
museum data. \emph{Biological Conservation}, 83(1), 9-17.
\doi{10.1016/S0006-3207(97)00048-7}
}
