% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrarabdotos}
\alias{metrarabdotos}
\title{\emph{Metrarabdotos} n. sp. 5 sightings}
\format{
A numeric vector with 11 values.
}
\usage{
metrarabdotos
}
\description{
Stratigraphic intervals for \emph{Metrarabdotos} n. sp. 5 from Marshall (1994)
based on data supplied by A. H. Cheetham.
}
\note{
The oldest sightings are set at time 0.
}
\references{
\strong{Key Reference}

Marshall, C. R. (1994). Confidence intervals on stratigraphic ranges:
partial relaxation of the assumption of randomly distributed fossil
horizons. \emph{Paleobiology}, 20(4), 459-469. \doi{10.1017/S0094837300012938}
}
\keyword{datasets}
