% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDisNet.R
\name{multicypher}
\alias{multicypher}
\title{Multiple statements execute}
\usage{
multicypher(
  graph,
  queries,
  parameters = NULL,
  result = c("row", "graph"),
  arraysAsStrings = TRUE,
  eltSep = " || "
)
}
\arguments{
\item{graph}{the neo4j connection}

\item{parameters}{parameters for the cypher query}

\item{result}{the way to return results. "row" will return a data frame
and "graph" will return a list of nodes, a list of relationships and
a list of paths (vectors of relationships identifiers).}

\item{eltSep}{if result="row" and arraysAsStrings is TRUE (default) array from
neo4j are converted to strings and array elementes are separated by eltSep.}

\item{statements}{cypher query}

\item{assayAsStrings}{if result="row" and arraysAsStrings is TRUE (default)
array from neo4j are converted to strings and array elements are separated by eltSep.}
}
\description{
Multiple statements execute
}
