% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_disNet.R
\name{normalize_disNet}
\alias{normalize_disNet}
\title{Normalize disease network}
\usage{
normalize_disNet(x)
}
\arguments{
\item{disNet}{a disease network to normalize}
}
\value{
A normalized disease network, disNet
}
\description{
This function ensure that disease network properties are fulfilled.
}
