% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDisNet.R, R/normalizeDisNet.R
\name{normalizeDisNet}
\alias{normalizeDisNet}
\title{Normalize disease network}
\usage{
normalizeDisNet(disNet)

normalizeDisNet(disNet)
}
\arguments{
\item{disNet}{a disease network to normalize}
}
\value{
A normalized disease network, disNet

A normalized disease network, disNet
}
\description{
This function ensure that disease network properties are fulfilled.

This function ensure that disease network properties are fulfilled.
}
