% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orchid}
\alias{orchid}
\title{Mauritian orchid sightings}
\format{
A numeric vector with 11 values.
}
\usage{
orchid
}
\description{
Mauritian orchid (\emph{Angraecum aff. rutenbergianum}) sighting records from
Solow & Roberts (2003).
}
\references{
\strong{Key Reference}

Solow, A. R., & Roberts, D. L. (2003). A nonparametric test for extinction
based on a sighting record. \emph{Ecology}, 84(5), 1329-1332.
\doi{10.1890/0012-9658(2003)084[1329:ANTFEB]2.0.CO;2}
}
\keyword{datasets}
