% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-utilities.R
\name{plot_relations}
\alias{plot_relations}
\title{Identify and show all relationships}
\usage{
plot_relations(
  ids,
  relationship = c("xref"),
  direct = F,
  sep = ":",
  verbose = FALSE
)
}
\arguments{
\item{ids}{String. A vector with identifier to convert formatted as DB:id (eg. "MONDO:0005027)}

\item{relationship}{String. Type of relationship to extract ("xref", "parent", "child",
"phenotype", "disease", "alternative"). Default: xref.}

\item{direct}{Boolean. Only getting direct relationships or moving through the disease
network to return all connected relationship.}

\item{sep}{String. Separator used, default: ":".}

\item{verbose}{show query input (default = FALSE)}
}
\description{
Takes the datatable that was obtained from \link{get_network} as input and
visualizes the nodes and edges as a visnetwork
}
