% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_disNet.R
\name{print.disNet}
\alias{print.disNet}
\title{Prints a disNet object}
\usage{
\method{print}{disNet}(x, ...)
}
\arguments{
\item{x}{disNet object}

\item{...}{additional parameters}
}
\description{
Prints a disNet object
}
