% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_disNet.R
\name{print.setDisNet}
\alias{print.setDisNet}
\title{Prints a setDisNet object}
\usage{
\method{print}{setDisNet}(x, ...)
}
\arguments{
\item{x}{disNet object}

\item{...}{additional parameters}
}
\description{
Prints a setDisNet object
}
