% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_concept.R
\name{process_cql}
\alias{process_cql}
\title{helper function to send the query and format the output}
\usage{
process_cql(cql = cql, ids = ids, to_db = to_db, verbose = verbose)
}
\arguments{
\item{cql}{String. Cql query to send}

\item{ids}{String. Vector of identifiers}

\item{to_db}{String. database to filter for.}

\item{verbose}{Boolean. verbose output.}
}
\description{
Prepares, send and return the output of the query
}
