% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RO09B1.R
\name{RO09B1}
\alias{RO09B1}
\title{Rout et al.'s (2009) "Declining" model}
\usage{
RO09B1(records)
}
\arguments{
\item{records}{sighting records in \code{cbin} format (see
\code{\link{convert_dodo}} for details).}
}
\value{
a \code{list} object with the original parameters and the rate of
decline (\eqn{a}) and Bayes factor included as elements.
}
\description{
Equation 6 and others from Rout et al. 2009. Estimates a pre-extinction rate
of decline, and a Bayes factor comparing competing hypotheses of extinction
/ persistence.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\examples{
# Run an example analysis using the Slender-billed Curlew data
\dontrun{
RO09B1(curlew$cbin)
}

}
\references{
\strong{Key Reference}

Rout, T. M., Salomon, Y., & McCarthy, M. A. (2009). Using sighting records
to declare eradication of an invasive species. \emph{Journal of Applied Ecology},
46(1), 110-117. \doi{10.1111/j.1365-2664.2008.01586.x}

\strong{Other References}

Solow, A. R. (1993). Inferring Extinction from Sighting Data. \emph{Ecology},
74(3), 962-964. \doi{10.2307/1940821}
}
