% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS03F1.R
\name{RS03F1}
\alias{RS03F1}
\title{Roberts & Solow's (2003) "Optimal Linear Estimation" model}
\usage{
RS03F1(
  records,
  alpha = 0.05,
  conf.int = "two-sided",
  k = ifelse(length(records) >= 10, 10, length(records)),
  test.time = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{records}{sighting records in \code{ccon} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}

\item{conf.int}{what kind of confidence interval to present. Valid options
are \code{"two-sided"} (the default) or \code{"one-sided"}.}

\item{k}{number of most recent sighting records to use (defaults to 10, or
the whole sighting record, if there are fewer than 10 records).}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}
}
\value{
a \code{list} object with the original parameters and the p-value, point
estimate, and confidence interval included as elements. The confidence
interval is a two-element numeric vector called \code{conf.int}.
}
\description{
Equations from Roberts & Solow 2003 (Equations 12-18 from Solow 2005).
Estimates a p-value for testing competing hypotheses of
extinction/non-extinction, and a \eqn{1 - \alpha} confidence interval and
point estimate on the time of extinction.
}
\note{
All sighting records are assumed to be certain.
}
\examples{
# Run the Dodo analysis from Roberts & Solow 2003
RS03F1(dodos, test.time = 2002)
# Run an example analysis using the Slender-billed Curlew data
\dontrun{
RS03F1(curlew$ccon, test.time = 2022)
}

}
\references{
\strong{Key Reference}

Roberts, D. L., & Solow, A. R. (2003). Flightless birds: when did the dodo
become extinct? \emph{Nature}, 426(6964), 245. \doi{10.1038/426245a}

\strong{Other References}

Solow, A. R. (2005). Inferring extinction from a sighting record.
\emph{Mathematical Biosciences}, 195(1), 47-55. \doi{10.1016/j.mbs.2005.02.001}

Rivadeneira, M. M., Hunt, G., & Roy, K. (2009). The use of sighting records
to infer species extinctions: an evaluation of different methods. \emph{Ecology},
90(5), 1291-1300. \doi{10.1890/08-0316.1}

Clements, C. F., Collen, B., Blackburn, T. M., & Petchey, O. L. (2014).
Effects of recent environmental change on accuracy of inferences of
extinction status. \emph{Conservation Biology}, 28(4), 971-981.
\doi{10.1111/cobi.12329}
}
