% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SB14B2.R
\name{SB14B2}
\alias{SB14B2}
\title{Solow & Beet's (2014) "Model 2" model}
\usage{
SB14B2(records, init.time, increment = 0.01)
}
\arguments{
\item{records}{sighting records in \code{ubin} format (see
\code{\link{convert_dodo}} for details).}

\item{init.time}{start of the observation period.}

\item{increment}{step size used for integration. Defaults to 0.01, following
the original code from Solow & Beet 2014.}
}
\value{
a \code{list} object with the original parameters and the Bayes factor
included as elements.
}
\description{
The second model from Solow & Beet 2014. Estimates a Bayes factor comparing
competing hypotheses of extinction / persistence.
}
\note{
Sampling effort is assumed to be constant. Uses a uniform prior.
}
\examples{
# Run the Ivory-billed Woodpecker analysis from Solow & Beet 2014
SB14B2(records = woodpecker$ubin, init.time = 1897, increment = 0.01)
# Run an example analysis using the Slender-billed Curlew data
\dontrun{
SB14B2(curlew$ubin, init.time = 1817, increment = 0.01)
}

}
\references{
\strong{Key Reference}

Solow, A. R., & Beet, A. R. (2014). On uncertain sightings and inference
about extinction. \emph{Conservation Biology}, 28(4), 1119-1123.
\doi{10.1111/cobi.12309}

\strong{Other References}

Carlson, C. J., Bond, A. L., & Burgio, K. R. (2018). Estimating the
extinction date of the thylacine with mixed certainty data.
\emph{Conservation Biology}, 32(2), 477-483. \doi{10.1111/cobi.13037}

Carlson, C. J., Burgio, K. R., Dallas, T. A., & Bond, A. L. (2018). Spatial
extinction date estimation: a novel method for reconstructing spatiotemporal
patterns of extinction and identifying potential zones of rediscovery.
\emph{bioRxiv preprint}. \doi{10.1101/279679}

Kodikara, S., Demirhan, H., & Stone, L. (2018). Inferring about the
extinction of a species using certain and uncertain sightings.
\emph{Journal of Theoretical Biology}, 442, 98-109.
\doi{10.1016/j.jtbi.2018.01.015}

Burgio, K. R., Carlson, C. J., Bond, A. L., Rubega, M. A., & Tingley, M. W.
(2021). The two extinctions of the Carolina Parakeet
\emph{Conuropsis carolinensis}. \emph{Bird Conservation International}, 1-8.
\doi{10.1017/s0959270921000241}
}
\seealso{
\code{\link[=SB14B1]{SB14B1()}}]
}
