% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-utilities.R
\name{set_dodo_version}
\alias{set_dodo_version}
\title{Feeding DODO: Set the DODO version}
\usage{
set_dodo_version(dodoInstance, dodoVersion)
}
\arguments{
\item{dodoInstance}{instance of DODO to be set}

\item{dodoVersion}{version of DODO to be set}
}
\description{
Not exported to avoid unintended modifications of the DB.
This function is used when modifying the DODO content.
}
