% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-utilities.R
\name{set_labels}
\alias{set_labels}
\title{Setting node label in DODO}
\usage{
set_labels(concept, nodes)
}
\arguments{
\item{concept}{concept type}

\item{nodes}{nodes to set label for}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
