% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff_disNet.R
\name{setdiff_disNet}
\alias{setdiff_disNet}
\title{Substract one disease network from another}
\usage{
setdiff_disNet(disNet1, disNet2)
}
\arguments{
\item{disNet1}{the original disease network}

\item{disNet2}{the disease network to substract}
}
\value{
A normalized disease network
}
\description{
Splitting a disNet object into multiple subsets based the provided list
containing vectors of disease identifiers. This generates a new type of S3 object: \emph{setDisNet} object.
Note that only the provided disease identifiers will be included into the setDisNet objects, all others
are removed from the object.
}
