% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff_disNet.R
\name{setdiffDisNet}
\alias{setdiffDisNet}
\title{Substract one disease network from another}
\usage{
setdiffDisNet(disNet1, disNet2)
}
\arguments{
\item{disNet1}{the original disease network}

\item{disNet2}{the disease network to substract}

\item{keepAllNodes}{if FALSE (default) all nodes of disNet2 are substracted.
If TRUE all nodes of disNet1 are kept even if not involved in final children
or xref relationships.}

\item{keepAllSeed}{if TRUE (default) all seed are kept in nodes after
filtering.}
}
\value{
A normalized disease network
}
\description{
Splitting a disNet object into multiple subsets based the provided list
containing vectors of disease identifiers. This generates a new type of S3 object: \emph{setDisNet} object.
Note that only the provided disease identifiers will be included into the setDisNet objects, all others
are removed from the object.
}
