% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_concept.R
\name{shinyConcept}
\alias{shinyConcept}
\title{##################################################################################
Shiny app to identify and browse indications}
\usage{
shinyConcept()
}
\description{
##################################################################################
Shiny app to identify and browse indications
}
