% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_disNet.R
\name{slice_disNet}
\alias{slice_disNet}
\title{Slice a disNet}
\usage{
slice_disNet(x, n)
}
\arguments{
\item{x}{disNet object}

\item{n}{Integer row values. Provide either positive values to keep, or
negative values to drop. The values provided must be either all positive
or all negative. Indices beyond the number of rows in the input are silently ignored.}
}
\description{
Choose rows by their ordinal position in the disNet
}
