% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO16B2.R
\name{SO16B2}
\alias{SO16B2}
\title{Solow's (2016) "Sequential" model}
\usage{
SO16B2(
  records,
  init.time = min(records),
  test.time,
  curr.time = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{records}{numeric vector object containing all sighting records of the
taxon of interest.}

\item{init.time}{start of the observation period. Defaults to the time of
the first sighting, in which case this sighting is removed from the record.}

\item{test.time}{time point to sequentially calculate extinction
probability at. Must not be earlier than the time of the most recent
sighting nor later than \code{curr.time}.}

\item{curr.time}{end of the observation period, typically the present day
(defaults to the current year).}
}
\value{
a \code{list} object with the original parameters and p(extant) included
as elements.
}
\description{
Equation 9 from Solow 2016. Estimates a posterior probability that the
species is extant.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant. Uses the uniform prior (1 / (m + T)) specified on p. 797.
}
\examples{
# Run the Dodo analysis from Solow 2016
SO16B2(dodos, test.time = 1672, curr.time = 2015)
# Note that Solow 2016 presents p(extinct), which is 1 - p(extant) used here.
# CHECK

}
\references{
\strong{Key Reference}

Solow, A. R. (2016). A simple Bayesian method of inferring extinction:
comment. \emph{Ecology}, 97(3), 796-798. \doi{10.1890/15-0336.1}

\strong{Other References}

Alroy, J. (2014). A simple Bayesian method of inferring extinction.
\emph{Paleobiology}, 40(4), 584-607. \doi{10.1666/13074}

Alroy, J. (2016). A simple Bayesian method of inferring extinction: reply.
\emph{Ecology}, 97(3), 798-800. \doi{10.1002/ecy.1321}
}
\seealso{
\code{\link[=AL14B1]{AL14B1()}}; \code{\link[=SO16B1]{SO16B1()}}
}
