% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO93B1.R
\name{SO93B1}
\alias{SO93B1}
\title{Solow's (1993) "Bayesian" model}
\usage{
SO93B1(
  records,
  init.time = min(records),
  test.time = as.numeric(format(Sys.Date(), "\%Y")),
  pi = 0.5
)
}
\arguments{
\item{records}{sighting records in \code{ccon} format (see
\code{\link{convert_dodo}} for details).}

\item{init.time}{start of the observation period. Defaults to the time of
the first sighting, in which case this sighting is removed from the record.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}

\item{pi}{prior probability that \eqn{H_0} is true (defaults to
\eqn{\pi = 0.5}).}
}
\value{
a \code{list} object with the original parameters and the Bayes factor
and p(extant) included as elements.
}
\description{
Equation 3 and others from Solow 1993. Estimates a Bayes factor and
posterior probability that the species is extant at the test time.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\examples{
# Run the Caribbean Monk Seal analysis from Solow 1993
SO93B1(monk_seal, test.time = 1992)
# Run an example analysis using the Slender-billed Curlew data
\dontrun{
SO93B1(curlew$ccon, init.time = 1817, test.time = 2022)
}

}
\references{
\strong{Key Reference}

Solow, A. R. (1993). Inferring Extinction from Sighting Data. \emph{Ecology},
74(3), 962-964. \doi{10.2307/1940821}
}
\seealso{
\code{\link[=SO93F1]{SO93F1()}}
}
