% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO93F1.R
\name{SO93F1}
\alias{SO93F1}
\title{Solow's (1993) "Classical" model}
\usage{
SO93F1(
  records,
  alpha = 0.05,
  init.time = min(records),
  test.time = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{records}{sighting records in \code{ccon} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}

\item{init.time}{start of the observation period. Defaults to the time of
the first sighting, in which case this sighting is removed from the record.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}
}
\value{
a \code{list} object with the original parameters and the p-value, point
estimate, and confidence interval included as elements. The confidence
interval is a two-element numeric vector called \code{conf.int}.
}
\description{
Equation 2 from Solow 1993 (Equation 3 from Solow 2005), and Equations 4
and 5 from Solow 2005. Estimates a p-value for testing competing hypotheses
of extinction/non-extinction, and a one-tailed \eqn{1 - \alpha} confidence
interval and point estimate on the time of extinction.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\examples{
# Run the Caribbean Monk Seal analysis from Solow 1993
SO93F1(monk_seal, test.time = 1992)
# Run an example analysis using the Slender-billed Curlew data
SO93F1(curlew$ccon, init.time = 1817, test.time = 2022)

}
\references{
\strong{Key Reference}

Solow, A. R. (1993). Inferring Extinction from Sighting Data. \emph{Ecology},
74(3), 962-964. \doi{10.2307/1940821}

\strong{Other References}

Solow, A., & Helser, T. (2000). Detecting extinction in sighting data. In S.
Ferson & M. Burgman (Eds.), \emph{Quantitative methods for conservation biology}
(pp. 1-6). Springer-Verlag. \doi{10.1007/b97704}

Solow, A. R. (2005). Inferring extinction from a sighting record.
\emph{Mathematical Biosciences}, 195(1), 47-55. \doi{10.1016/j.mbs.2005.02.001}

Rivadeneira, M. M., Hunt, G., & Roy, K. (2009). The use of sighting records
to infer species extinctions: an evaluation of different methods. \emph{Ecology},
90(5), 1291-1300. \doi{10.1890/08-0316.1}
}
\seealso{
\code{\link[=SO93B1]{SO93B1()}}
}
