% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO93F2.R
\name{SO93F2}
\alias{SO93F2}
\title{Solow's (1993) "Declining" Model}
\usage{
SO93F2(
  records,
  alpha = 0.05,
  init.time = min(records),
  test.time = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{records}{numeric vector object containing all sighting records of the
taxon of interest.}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}

\item{init.time}{start of the observation period. Defaults to the time of
the first sighting, in which case  this sighting is removed from the record.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}
}
\value{
a \code{list} object with the original parameters and the p-value, point
estimate, and confidence interval included as elements. The confidence
interval is a two-element numeric vector called \code{conf.int}.
}
\description{
Equations 3 and 4 from Solow 1993 (Equation 7 from Solow 2005), and
Equation 8 from Solow 2005. Estimates a p-value for testing competing
hypotheses of extinction/non-extinction, and a one-tailed \eqn{1 - \alpha}
confidence interval and point estimate on the time of extinction.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\references{
\strong{Key Reference}

Solow, A. R. (1993). Inferring Extinction in a Declining Population.
\emph{Journal of Mathematical Biology}, 32(1), 79-82. \doi{10.1007/Bf00160376}

\strong{Other References}

Solow, A. R. (2005). Inferring extinction from a sighting record.
\emph{Mathematical Biosciences}, 195(1), 47-55. \doi{10.1016/j.mbs.2005.02.001}
}
