% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_disNet.R
\name{split_disNet}
\alias{split_disNet}
\title{Split a disNet}
\usage{
split_disNet(disNet, diseaseList)
}
\arguments{
\item{disNet}{the disease network to split}

\item{diseaseList}{list character vectors of disease ids (format "MONDO:0005027")}
}
\value{
A a list of normalized disease networks
}
\description{
Splitting a disNet into a list of disNets based on a list of vectors containing disease identifiers. Generates a S3 object
\emph{setDisNet}.
}
