% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR03F1.R
\name{SR03F1}
\alias{SR03F1}
\title{Solow & Roberts' (2003) "Non-parametric" model}
\usage{
SR03F1(
  records,
  alpha = 0.05,
  test.time = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{records}{sighting records in \code{ccon} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}
}
\value{
a \code{list} object with the original parameters and the p-value, point
estimate, and confidence interval included as elements. The confidence
interval is a two-element numeric vector called \code{conf.int}.
}
\description{
Equation 4 from Solow & Roberts 2003 and Equations 9-11 from Solow 2005.
Estimates a p-value for testing competing hypotheses of
extinction/non-extinction, and a one-sided \eqn{1 - \alpha} confidence
interval and point estimate on the time of extinction.
}
\note{
All sighting records are assumed to be certain.
}
\examples{
# Run the Mauritian orchid analysis from Solow & Roberts 2003
SR03F1(orchid, test.time = 1996)
\dontrun{
# Run an example analysis using the Slender-billed Curlew data
SR03F1(curlew$ccon, test.time = 2022)
}

}
\references{
\strong{Key Reference}

Solow, A. R., & Roberts, D. L. (2003). A nonparametric test for extinction
based on a sighting record. \emph{Ecology}, 84(5), 1329-1332.
\doi{10.1890/0012-9658(2003)084[1329:ANTFEB]2.0.CO;2}

\strong{Other References}

Robson, D. S., & Whitlock, J. H. (1964). Estimation of a truncation point.
\emph{Biometrika}, 51(1-2), 33-39. \doi{10.1093/biomet/51.1-2.33}

Solow, A. R. (2003). Estimation of stratigraphic ranges when fossil finds
are not randomly distributed. \emph{Paleobiology}, 29(2), 181-185.
\doi{10.1666/0094-8373(2003)029<0181:EOSRWF>2.0.CO;2}

Solow, A. R. (2005). Inferring extinction from a sighting record.
\emph{Mathematical Biosciences}, 195(1), 47-55. \doi{10.1016/j.mbs.2005.02.001}

Rivadeneira, M. M., Hunt, G., & Roy, K. (2009). The use of sighting records
to infer species extinctions: an evaluation of different methods. \emph{Ecology},
90(5), 1291-1300. \doi{10.1890/08-0316.1}
}
