% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS89B1.R
\name{SS89B1}
\alias{SS89B1}
\title{Strauss & Sadler's (1989) "Bayesian" Model}
\usage{
SS89B1(
  records,
  alpha = 0.05,
  t.max = max(records) * 2,
  length.out = 1e+06,
  keep.dists = FALSE
)
}
\arguments{
\item{records}{numeric vector object containing all sighting records of the
taxon of interest.}

\item{alpha}{desired threshold level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} credible interval.}

\item{t.max}{maximum time to estimate posterior density out to (defaults to
double the most recent record time).}

\item{length.out}{number of posterior samples to generate (defaults to 1
million).}

\item{keep.dists}{whether to include the posterior distribution vectors in
the final function output or not (defaults to \code{FALSE}).}
}
\value{
a \code{list} object with the original parameters and the point estimate
and credible interval included as elements. The credible interval is a
two-element numeric vector called \code{cred.int}. The posterior distribution
vectors are also included, if \code{keep.dists} is set to \code{TRUE}.
}
\description{
Equation 26 from Strauss & Sadler (1989), assuming the prior distribution
from equation 22. Estimates a posterior distribution on time of extinction,
with associated point estimate and one-sided credible interval.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\references{
\strong{Key Reference}

Strauss, D., & Sadler, P. M. (1989). Classical Confidence Intervals and
Bayesian Probability Estimates for Ends of Local Taxon Ranges.
\emph{Mathematical Geology}, 21(4), 411-421. \doi{10.1007/Bf00897326}
}
