% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS89F1.R
\name{SS89F1}
\alias{SS89F1}
\title{Strauss & Sadler's (1989) "Classical" model}
\usage{
SS89F1(records, alpha = 0.05)
}
\arguments{
\item{records}{sighting records in \code{ccon} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}
}
\value{
a \code{list} object with the original parameters and the point estimate
and confidence interval included as elements. The confidence interval is a
two-element numeric vector called \code{conf.int}.
}
\description{
Equations 8 and 20 from Strauss & Sadler 1989. Estimates a one-tailed
\eqn{1 - \alpha} confidence interval and point estimate on the time of
extinction.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\examples{
# Run an example analysis using the Caribbean Monk Seal data
SS89F1(monk_seal)
# Run an example analysis using the Slender-billed Curlew data
\dontrun{SS89F1(curlew$ccon)}

}
\references{
\strong{Key Reference}

Strauss, D., & Sadler, P. M. (1989). Classical Confidence Intervals and
Bayesian Probability Estimates for Ends of Local Taxon Ranges. Mathematical
Geology, 21(4), 411-421. \doi{10.1007/Bf00897326}

\strong{Other References}

Marshall, C. R. (1990). Confidence intervals on stratigraphic ranges.
\emph{Paleobiology}, 16(1), 1-10. \doi{10.1017/S0094837300009672}

McFarlane, D. A. (1999). A Comparison of Methods for the Probabilistic
Determination of Vertebrate Extinction Chronologies. In R. D. E. MacPhee
(Ed.), \emph{Extinctions in Near Time} (pp. 95-103). Springer US.
\doi{10.1007/978-1-4757-5202-1_5}

Rivadeneira, M. M., Hunt, G., & Roy, K. (2009). The use of sighting records
to infer species extinctions: an evaluation of different methods. \emph{Ecology},
90(5), 1291-1300. \doi{10.1890/08-0316.1}
}
\seealso{
\code{\link[=SS89B1]{SS89B1()}}
}
