% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starling}
\alias{starling}
\title{Pohnpei Mountain Starling sightings}
\format{
a \code{data.frame} with 5 columns and 84 rows.
}
\usage{
starling
}
\description{
Pohnpei Mountain Starling (\emph{Aplonis pelzelni}) sighting record from Lee
(2014). The temporal range (\code{init.time} to \code{test.time}) of these data
is 1930 to 2013. Format: \code{iucn}.
}
\references{
\strong{Key Reference}

Lee, T. E. (2014). A simple numerical tool to infer whether a species is
extinct. \emph{Methods in Ecology and Evolution}, 5(8), 791-796.
\doi{10.1111/2041-210x.12227}
}
\keyword{datasets}
