% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH13B1.R
\name{TH13B1}
\alias{TH13B1}
\title{Thompson et al.'s (2013) "Deterministic" model}
\usage{
TH13B1(records, priors, certain = c(1), PXT = NULL, PE = NULL)
}
\arguments{
\item{records}{sighting records in \code{umcd} format (see
\code{\link{convert_dodo}} for details).}

\item{priors}{nested \code{list} with two elements: \code{p} and \code{q}, each of which is
a \code{list} with the same number of elements as there are sighting classes. For
this deterministic model, each element is the point prior for the rate
\eqn{p_i^\alpha} or \eqn{q_i^\alpha}.}

\item{certain}{which sighting types are considered certain. Defaults to
\code{c(1)} (i.e. only the first sighting class is certain).}

\item{PXT}{estimate of \eqn{P(X_T)}. Defaults to \code{NULL}, in which case
\eqn{P(X_T) = TN / T}.}

\item{PE}{estimate of \eqn{P_E}. Defaults to \code{NULL}, in which case
\eqn{P_E = 1 / T}.}
}
\value{
a \code{list} object with the original parameters and p(extant) included
as elements.
}
\description{
The deterministic variant (\eqn{P_D(X_T|\bm{s})}) of Equation 9 in Thompson
et al. 2013. Estimates a posterior probability that the species is extant at
the end of the survey period.
}
\note{
Sampling effort is assumed to be constant.
}
\examples{
# Run the example analysis from Thompson et al. 2013 (Table 1 etc.)
TH13B1(
  records = thompson_table1,
  priors = list(
    p = list(p1 = 0.30, p2 = 0.50, p3 = 0.45),
    q = list(q1 = 1.00, q2 = 0.45, q3 = 0.30)
  ),
  certain = c(1)
)
# Run an example analysis using the Slender-billed Curlew data
\dontrun{
TH13B1(
  records = curlew$umcd,
  priors = list(
    p = list(p1 = 0.5, p2 = 0.5, p3 = 0.5, p4 = 0.5, p5 = 0.5, p6 = 0.5),
    q = list(q1 = 1.0, q2 = 1.0, q3 = 0.5, q4 = 0.5, q5 = 0.5, q6 = 0.5)
  ),
  certain = c(1, 2)
)
}

}
\references{
\strong{Key Reference}

Thompson, C. J., Lee, T. E., Stone, L., McCarthy, M. A., & Burgman, M. A.
(2013). Inferring extinction risks from sighting records.
\emph{Journal of Theoretical Biology}, 338, 16-22.
\doi{10.1016/j.jtbi.2013.08.023}
}
\seealso{
\code{\link[=TH13B2]{TH13B2()}}; \code{\link[=TH13B3]{TH13B3()}}
}
