% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH13B2.R
\name{TH13B2}
\alias{TH13B2}
\title{Thompson et al.'s (2013) "Quenched" model}
\usage{
TH13B2(records, priors, certain = 1, PXT = NULL, PE = NULL, n.iter = 1e+05)
}
\arguments{
\item{records}{sighting records in \code{umcd} format (see
\code{\link{convert_dodo}} for details).}

\item{priors}{nested \code{list} with two elements: \code{p} and \code{q}, each of which is
a \code{list} with the same number of elements as there are sighting classes. For
the quenched model, each element is the a two-element vector with the lower
and upper bounds of a uniform prior for the rate \eqn{p_i^\alpha} or
\eqn{q_i^\alpha}.}

\item{certain}{which sighting types are considered certain. Defaults to
\code{c(1)} (i.e. only the first sighting class is certain).}

\item{PXT}{estimate of \eqn{P(X_T)}. Defaults to \code{NULL}, in which case
\eqn{P(X_T) = TN / T}.}

\item{PE}{estimate of \eqn{P_E}. Defaults to \code{NULL}, in which case
\eqn{P_E = 1 / T}.}

\item{n.iter}{number of iterations to calculate averages over. Defaults to
100,000, which is usually sufficient to ensure accurate estimates.}
}
\value{
a \code{list} object with the original parameters and p(extant) included
as elements.
}
\description{
The quenched variant (\eqn{P_Q(X_T|\bm{s})}) of Equation 9 in Thompson et
al. 2013. Estimates a posterior probability that the species is extant at
the end of the survey period.
}
\note{
Sampling effort is assumed to be constant.
}
\examples{
# Run the example analysis from Thompson et al. 2013 (Table 1 etc.)
TH13B2(
  records = thompson_table1,
  priors = list(
    p = list(p1 = c(0.2, 0.4), p2 = c(0.4, 0.6), p3 = c(0.3, 0.6)),
    q = list(q1 = c(1.0, 1.0), q2 = c(0.2, 0.7), q3 = c(0.1, 0.5))
  ),
  certain = c(1),
  n.iter = 1e3
)
# Run an example analysis using the Slender-billed Curlew data
TH13B2(
  records = curlew$umcd,
  priors = list(
    p = list(p1 = c(0, 1), p2 = c(0, 1), p3 = c(0, 1),
             p4 = c(0, 1), p5 = c(0, 1), p6 = c(0, 1)),
    q = list(q1 = c(1.0, 1.0), q2 = c(1.0, 1.0), q3 = c(0.0, 1.0),
             q4 = c(0.0, 1.0), q5 = c(0.0, 1.0), q6 = c(0.0, 1.0))),
  certain = c(1, 2),
  n.iter = 1e3
)

}
\references{
\strong{Key Reference}

Thompson, C. J., Lee, T. E., Stone, L., McCarthy, M. A., & Burgman, M. A.
(2013). Inferring extinction risks from sighting records.
\emph{Journal of Theoretical Biology}, 338, 16-22.
\doi{10.1016/j.jtbi.2013.08.023}
}
\seealso{
\code{\link[=TH13B1]{TH13B1()}}; \code{\link[=TH13B3]{TH13B3()}}
}
