% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH17I1.R
\name{TH17I1}
\alias{TH17I1}
\title{Thompson et al.'s (2017) "Iterative" model}
\usage{
TH17I1(records, surveys)
}
\arguments{
\item{records}{sighting records in \code{iucn} format (see
\code{\link{convert_dodo}} for details).}

\item{surveys}{survey effort data (incorporating both dedicated and passive
survey effort). A \code{data.frame} with 11 columns: \code{time}, \code{survey}, \code{epsilon},
\code{epsilon_lower}, \code{epsilon_upper}, \code{p_i}, \code{p_i_lower}, \code{p_i_upper}, \code{p_r},
\code{p_r_lower}, and \code{p_r_upper}. The number of rows must match \code{records}.}
}
\value{
a \code{list} object with the original parameters and the iterative and
}
\description{
The records and surveys model from Thompson et al. 2017. Estimates a
probability that the species is extant at the end of the observation period.
}
\note{
This model incorporates both sighting uncertainty and variable survey effort.
}
\examples{
# Run the Alaotra Grebe analysis from Thompson et al. 2017
TH17I1(records = grebe, surveys = grebe_surveys)

}
\references{
\strong{Key Reference}

Thompson, C. J., Koshkina, V., Burgman, M. A., Butchart, S. H. M., & Stone,
L. (2017). Inferring extinctions II: A practical, iterative model based on
records and surveys. \emph{Biological Conservation}, 214, 328-335.
\doi{10.1016/j.biocon.2017.07.029}
}
