% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TH19B1.R
\name{TH19B1}
\alias{TH19B1}
\title{Thompson et al.'s (2019) "Bayesian Updating" model}
\usage{
TH19B1(records, surveys, init.time, test.time, pi = 0.5)
}
\arguments{
\item{records}{sighting records in \code{iucn} format (see
\code{\link{convert_dodo}} for details).}

\item{surveys}{survey effort data (incorporating both dedicated and passive
survey effort). A \code{data.frame} with 11 columns: \code{time}, \code{survey}, \code{epsilon},
\code{epsilon_lower}, \code{epsilon_upper}, \code{p_i}, \code{p_i_lower}, \code{p_i_upper}, \code{p_r},
\code{p_r_lower}, and \code{p_r_upper}. The number of rows must match \code{records}.}

\item{init.time}{the time point at which to start the Bayesian updating
algorithm. At this point, \eqn{P(X_t)} is set to \code{pi}.}

\item{test.time}{time point to retrospectively calculate extinction
probability at.}

\item{pi}{prior probability that the species is extant at \code{init.time}
(defaults to \eqn{\pi = 0.5}).}
}
\value{
a \code{list} object with the original parameters and the final
estimates of p(extant) (\eqn{P(X_t)}) and the cumulative Bayes Factor
included as elements.
}
\description{
The model from Thompson et al. 2019. Estimates a probability that the
species is extant at the test time, and a cumulative Bayes Factor at that
point.
}
\note{
This model incorporates both sighting uncertainty and variable survey effort.
}
\examples{
# Run the Alaotra Grebe analysis from Thompson et al. 2019 (Fig. 1)
TH19B1(
  records = grebe, surveys = grebe_surveys, init.time = 1990,
  test.time = 1998, pi = 0.5
)

}
\references{
\strong{Key Reference}

Thompson, C. J., Kodikara, S., Burgman, M. A., Demirhan, H., & Stone, L.
(2019). Bayesian updating to estimate extinction from sequential observation
data. \emph{Biological Conservation}, 229, 26-29.
\doi{10.1016/j.biocon.2018.11.003}
}
