% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thompson_table1}
\alias{thompson_table1}
\title{Example sightings from Thompson et al.}
\format{
a \code{data.frame} with 4 columns and 7 rows.
}
\usage{
thompson_table1
}
\description{
Example multi-class sighting record from Table 1 in Thompson et al. (2013).
The temporal range (\code{init.time} to \code{test.time}) of these data is 0 to 7.
Format: \code{umcd}.
}
\references{
\strong{Key Reference}

Thompson, C. J., Lee, T. E., Stone, L., McCarthy, M. A., & Burgman, M. A.
(2013). Inferring extinction risks from sighting records.
\emph{Journal of Theoretical Biology}, 338, 16-22.
\doi{10.1016/j.jtbi.2013.08.023}
}
\keyword{datasets}
