% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{verneuilinoides}
\alias{verneuilinoides}
\title{Verneuilinoides sp. A records}
\format{
a \code{numeric} vector with 36 values.
}
\usage{
verneuilinoides
}
\description{
Verneuilinoides sp. A (species 25) sighting record from Weiss & Marshall
(2014). Format: \code{cbin}.
}
\references{
\strong{Key Reference}

Weiss, R. E., & Marshall, C. R. (1999). The Uncertainty in the True End
Point of a Fossil's Stratigraphic Range When Stratigraphic Sections Are
Sampled Discretely. \emph{Mathematical Geology}, 31(4), 435-453.
\doi{10.1023/A:1007542725180}
}
\keyword{datasets}
