% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA16B1.R
\name{WA16B1}
\alias{WA16B1}
\title{Wang et al.'s (2016) "Adaptive" model}
\usage{
WA16B1(records, alpha = 0.05, init.time = min(records))
}
\arguments{
\item{records}{numeric vector object containing all sighting records of the
taxon of interest.}

\item{alpha}{desired threshold level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} credible interval.}

\item{init.time}{start of the observation period. Defaults to the time of
the first sighting, in which case this sighting is removed from the record.}
}
\value{
a \code{list} object with the original parameters and the point estimate
and credible interval included as elements. The credible interval is a
two-element numeric vector called \code{cred.int}.
}
\description{
The Adaptive Beta Method from Wang et al. 2016. Estimates a posterior
distribution on the time of extinction, with associated point estimate and
one-sided \eqn{1 - \alpha} credible interval.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\examples{
# Run the Anabarella analysis from Wang et al. 2016
WA16B1(-anabarella, alpha = 0.1)

}
\references{
\strong{Key Reference}

Wang, S. C., Everson, P. J., Zhou, H. J., Park, D., & Chudzicki, D. J.
(2016). Adaptive credible intervals on stratigraphic ranges when recovery
potential is unknown. \emph{Paleobiology}, 42(2), 240-256.
\doi{10.1017/pab.2015.37}
}
