% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weissmarshall_surveys}
\alias{weissmarshall_surveys}
\title{Weiss & Marshall surveys}
\format{
a \code{numeric} vector with 36 values.
}
\usage{
weissmarshall_surveys
}
\description{
Palaeontological survey data from Weiss & Marshall (2014).
}
\references{
\strong{Key Reference}

Weiss, R. E., & Marshall, C. R. (1999). The Uncertainty in the True End
Point of a Fossil's Stratigraphic Range When Stratigraphic Sections Are
Sampled Discretely. \emph{Mathematical Geology}, 31(4), 435-453.
\doi{10.1023/A:1007542725180}
}
\keyword{datasets}
