% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WM99B1.R
\name{WM99B1}
\alias{WM99B1}
\title{Weiss & Marshall's (1999) "Discrete-time" model}
\usage{
WM99B1(
  records,
  surveys,
  alpha = 0.05,
  init.time = min(surveys),
  test.time = max(surveys),
  priors
)
}
\arguments{
\item{records}{sighting records in \code{cbin} format (see
\code{\link{convert_dodo}} for details).}

\item{surveys}{a \code{numeric} vector of the survey times for each observation
in \code{records}.}

\item{alpha}{desired threshold level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} credible interval.}

\item{init.time}{start of the observation period.}

\item{test.time}{time point to retrospectively calculate extinction
probability at.}

\item{priors}{\code{list} with three elements: \code{lambda}, \code{c} and \code{d}. \code{lambda} is
the mean lifetime (half-life) for the exponential prior on \code{S}, the time of
extinction. \code{c} and \code{d} are the two shape parameters for the beta prior on
\code{pi}, the pre-extinction detection probability.}
}
\value{
a \code{list} object with the original parameters and the p(extant),
point estimate, and credible interval included as elements. The credible
interval is a two-element numeric vector called \code{cred.int}. The point
estimate is the median (not the mean) of the posterior distribution of
extinction time.
}
\description{
The model from Weiss & Marshall 1999. Estimates a posterior probability that
the species is extant at the test time, and a point estimate and one-tailed
\eqn{1 - \alpha} credible interval on the time of extinction.
}
\note{
All sighting records are assumed to be certain.
}
\examples{
# Run the Verneuilinoides sp. A analysis from Weiss & Marshall 1999
WM99B1(verneuilinoides, weissmarshall_surveys,
  priors = list(lambda = 800, c = 11, d = 70)
)
# Run the Eggerellina brevis analysis from Weiss & Marshall 1999
WM99B1(eggerellina_brevis, weissmarshall_surveys,
  priors = list(lambda = 800, c = 85, d = 17)
)
# Run an example analysis using the Slender-billed Curlew data
\dontrun{
WM99B1(curlew, 1817:2022, priors = list(lambda = 1e6, c = 1, d = 1))
}

}
\references{
\strong{Key Reference}

Weiss, R. E., & Marshall, C. R. (1999). The Uncertainty in the True End
Point of a Fossil's Stratigraphic Range When Stratigraphic Sections Are
Sampled Discretely. \emph{Mathematical Geology}, 31(4), 435-453.
\doi{10.1023/A:1007542725180}
}
