% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{woodpecker}
\alias{woodpecker}
\title{Ivory-billed Woodpecker sightings}
\format{
a \code{list} with 6 elements.
}
\usage{
woodpecker
}
\description{
Ivory-billed Woodpecker (\emph{Campephilus principalis}) sighting record from
Roberts et al. (2010), processed into various \code{dodo} formats. The temporal
range (\code{init.time} to \code{test.time}) of these data is 1897 to 2010. Formats:
\code{ccon}, \code{cbin}, \code{cdis}, \code{ucon}, \code{ubin}, \code{umcd}.
}
\references{
\strong{Key Reference}

Roberts, D. L., Elphick, C. S., & Reed, J. M. (2010). Identifying anomalous
reports of putatively extinct species and why it matters.
\emph{Conservation Biology}, 24(1), 189-196.
\doi{10.1111/j.1523-1739.2009.01292.x}
}
\keyword{datasets}
