% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{woodpecker1}
\alias{woodpecker1}
\title{Ivory-billed Woodpecker Sightings (Certain)}
\format{
A numeric vector with 22 values.
}
\usage{
woodpecker1
}
\description{
Ivory-billed Woodpecker (\emph{Campephilus principalis}) sighting records from
Roberts et al. (2010).
}
\note{
Only certain sightings are included.
}
\references{
\strong{Key Reference}

Roberts, D. L., Elphick, C. S., & Reed, J. M. (2010). Identifying anomalous
reports of putatively extinct species and why it matters. \emph{Conservation}
\emph{Biology}, 24(1), 189-196.\doi{10.1111/j.1523-1739.2009.01292.x}
}
\keyword{datasets}
