% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{woodpecker3}
\alias{woodpecker3}
\title{Ivory-billed Woodpecker sightings (variable certainty)}
\format{
A \code{data.frame} with 2 columns and 68 rows.
}
\usage{
woodpecker3
}
\description{
Ivory-billed Woodpecker (\emph{Campephilus principalis}) sighting records from
Roberts et al. (2010).
}
\note{
Sighting certainty values are the "extreme" values from Brook et al. 2019
(i.e. 0.99, 0.5, 0.01).
}
\references{
\strong{Key Reference}

Roberts, D. L., Elphick, C. S., & Reed, J. M. (2010). Identifying anomalous
reports of putatively extinct species and why it matters. \emph{Conservation}
\emph{Biology}, 24(1), 189-196.\doi{10.1111/j.1523-1739.2009.01292.x}

\strong{Other References}

Brook, B. W., Buettel, J. C., & Jaric, I. (2019). A fast re-sampling method
for using reliability ratings of sightings with extinction-date estimators.
\emph{Ecology}, 100(9), e02787. \doi{10.1002/ecy.2787}
}
\keyword{datasets}
