% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA26B2.R
\name{BA26B2}
\alias{BA26B2}
\title{Backstrom et al.'s (2026) "Uncertain No-effort" model}
\usage{
BA26B2(
  records,
  alpha = 0.05,
  init.time,
  test.time = init.time + nrow(records) - 1,
  priors = list(theta = c(1, 10), lambda_v = c(1, 1), lambda_i = c(1, 1)),
  n.chains = 4,
  n.iter = 110000,
  n.burnin = 10000,
  n.thin = 10
)
}
\arguments{
\item{records}{sighting records in \code{udis} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired threshold level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} credible interval.}

\item{init.time}{start of the observation period.}

\item{test.time}{time point to retrospectively calculate extinction
probability at. Defaults to the end of the observation period.}

\item{priors}{\code{list} with three elements: \code{theta}, \code{lambda_v} and \code{lambda_i},
themselves all \code{numeric} vectors of length two. The two elements in \code{theta}
are the shape parameters for the Beta hyperprior on \eqn{\theta}. They
default to (1, 10). The two elements in \code{lambda_v} and \code{lambda_i} are the
shape and rate parameters for the Gamma priors on \eqn{\lambda_v} and
\eqn{\lambda_i}. They both default to (1, 1).}

\item{n.chains}{number of MCMC chains to run. Defaults to 4.}

\item{n.iter}{number of iterations in each chain. Defaults to 110,000.}

\item{n.burnin}{number of iterations to discard as burn-in. Defaults to
10,000.}

\item{n.thin}{thinning rate. Defaults to 10.}
}
\value{
a \code{list} object with the original parameters and the p(extant),
point estimate, and credible interval included as elements. The credible
interval is a two-element numeric vector called \code{cred.int}.
}
\description{
Model 2 from Backstrom et al. 2026. Estimates a posterior probability that
the species is extant at the test time, and a point estimate and one-sided
\eqn{1 - \alpha} credible interval on the time of extinction.
}
\note{
Sampling effort is assumed to be constant.
}
\examples{
\dontrun{
# ADD
}

}
\references{
\strong{Key Reference}

Backstrom, L. J. et al. (in prep).
}
\seealso{
\code{\link[=BA26B1]{BA26B1()}}, \code{\link[=BA26B3]{BA26B3()}}, \code{\link[=BA26B4]{BA26B4()}}
}
