% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA26B3.R
\name{BA26B3}
\alias{BA26B3}
\title{Backstrom et al.'s (2026) "Certain-only Effort" model}
\usage{
BA26B3(
  records,
  effort,
  alpha = 0.05,
  init.time,
  test.time = init.time + length(records) - 1,
  priors = list(theta = c(1, 10), sigma = c(1)),
  n.chains = 4,
  n.iter = 110000,
  n.burnin = 10000,
  n.thin = 10
)
}
\arguments{
\item{records}{sighting records in \code{cdis} format (see
\code{\link{convert_dodo}} for details).}

\item{effort}{a \code{data.frame} of effort data, with the same number of rows as
\code{records} and as many columns as there are effort variables.}

\item{alpha}{desired threshold level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} credible interval.}

\item{init.time}{start of the observation period.}

\item{test.time}{time point to retrospectively calculate extinction
probability at. Defaults to the end of the observation period.}

\item{priors}{\code{list} with two elements: \code{theta} and \code{sigma}, themselves both
\code{numeric} vectors.  \code{theta} is of length two, with the two elements being the
shape parameters for the Beta hyperprior on \eqn{\theta}. They default to
(1, 10). \code{sigma} should either be of length one, or the same length as the
number of coefficients to estimate (i.e. \code{ncol(effort) + 1}). Defaults to 1.}

\item{n.chains}{number of MCMC chains to run. Defaults to 4.}

\item{n.iter}{number of iterations in each chain. Defaults to 110,000.}

\item{n.burnin}{number of iterations to discard as burn-in. Defaults to
10,000.}

\item{n.thin}{thinning rate. Defaults to 10.}
}
\value{
a \code{list} object with the original parameters and the p(extant),
point estimate, and credible interval included as elements. The credible
interval is a two-element numeric vector called \code{cred.int}.
}
\description{
Model 3 from Backstrom et al. 2026. Estimates a posterior probability that
the species is extant at the test time, and a point estimate and one-sided
\eqn{1 - \alpha} credible interval on the time of extinction.
}
\note{
All sighting records are assumed to be certain.
}
\examples{
\dontrun{
# ADD
}

}
\references{
\strong{Key Reference}

Backstrom, L. J. et al. (in prep).
}
\seealso{
\code{\link[=BA26B1]{BA26B1()}}, \code{\link[=BA26B2]{BA26B2()}}, \code{\link[=BA26B4]{BA26B4()}}
}
