% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BR19F1.R
\name{BR19F1}
\alias{BR19F1}
\title{Brook et al.'s (2019) "Solow" model}
\usage{
BR19F1(
  records,
  alpha = 0.05,
  init.time = min(records$time),
  test.time = as.numeric(format(Sys.Date(), "\%Y")),
  n.iter = 10000
)
}
\arguments{
\item{records}{sighting records in \code{ucon} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}

\item{init.time}{start of the observation period. Defaults to the time of
the first sighting, in which case this sighting is removed from the record.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}

\item{n.iter}{number of iterations to run (defaults to 10,000).}
}
\value{
a \code{list} object with the original parameters and the probability,
point estimate, and confidence interval included as elements. The confidence
interval is a two-element numeric vector called \code{conf.int}.
}
\description{
The Solow 1993a variant of the model from Brook et al. 2019. Estimates a
"probability" of extinction at the test time, and a two-sided
\eqn{1 - \alpha} confidence interval and point estimate on the time of
extinction. Sighting uncertainty is incorporated.
}
\note{
Sampling effort is assumed to be constant.
}
\examples{
# Run the "Extreme" Ivory-billed Woodpecker analysis from Brook et al. 2019
BR19F1(woodpecker$ucon, alpha = 0.1, test.time = 2010, n.iter = 1e3)
\dontrun{
# Run an example analysis using the Slender-billed Curlew data
BR19F1(curlew$ucon, test.time = 2022, n.iter = 1e3)
}

}
\references{
\strong{Key Reference}

Brook, B. W., Buettel, J. C., & Jaric, I. (2019). A fast re-sampling method
for using reliability ratings of sightings with extinction-date estimators.
\emph{Ecology}, 100(9), e02787. \doi{10.1002/ecy.2787}

\strong{Other References}

Solow, A. R. (1993). Inferring Extinction from Sighting Data. \emph{Ecology},
74(3), 962-964. \doi{10.2307/1940821}
}
\seealso{
\code{\link[=BR19F2]{BR19F2()}}, \code{\link[=BR19F3]{BR19F3()}}, \code{\link[=BR19F4]{BR19F4()}}, \code{\link[=BR19F5]{BR19F5()}}
}
