% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDisNet.R
\name{c.disNet}
\alias{c.disNet}
\title{Merge several disease networks}
\usage{
\method{c}{disNet}(..., list = NULL)
}
\arguments{
\item{...}{disease networks to merge}
}
\value{
A normalized merged disease network
}
\description{
Merge several disease networks
}
\examples{
disNet1 <- buildDisNetByTerm("epilep",fields = c("synonym","label","definition"))
disNet2 <- buildDisNetByTerm("alzheim",fields = c("synonym","label","definition"))
mergeDisNet <- mergeDisNet(disNet1,disNet2)

}
