% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB14B2.R
\name{CB14B2}
\alias{CB14B2}
\title{Caley & Barry's (2014) "Non-constant" model}
\usage{
CB14B2(
  records,
  alpha = 0.05,
  init.time = min(records),
  test.time = as.numeric(format(Sys.Date(), "\%Y")),
  burn.in = 1000,
  n.iter = 11000
)
}
\arguments{
\item{records}{numeric vector object containing all sighting records of the
taxon of interest.}

\item{alpha}{desired threshold level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} credible interval.}

\item{init.time}{start of the observation period. Defaults to the time of
the first sighting, in which case this sighting is removed from the record.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}

\item{burn.in}{number of initial iterations to discard as burn-in (defaults
to 1000).}

\item{n.iter}{number of iterations to run (defaults to 11000).}
}
\value{
a \code{list} object with the original parameters and the p(extant),
point estimate, and credible interval included as elements. The credible
interval is a two-element numeric vector called \code{cred.int}.
}
\description{
Non-constant population model from Caley & Barry 2014. Estimates a posterior
probability that the species is extant at the test time, and a point
estimate and one-tailed \eqn{1 - \alpha} credibile interval on the time of
extinction.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\references{
\strong{Key Reference}

Caley, P., & Barry, S. C. (2014). Quantifying extinction probabilities from
sighting records: inference and uncertainties. \emph{PLoS One}, 9(4), e95857.
\doi{10.1371/journal.pone.0095857}
}
\seealso{
\code{\link[=CB14B1]{CB14B1()}}
}
