% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_disNet.R
\name{cluster_disNet}
\alias{cluster_disNet}
\title{Cluster a disNet}
\usage{
cluster_disNet(
  disNet,
  clusterOn = c("xref"),
  forwardAmbiguity = 10000,
  backwardAmbiguity = 1,
  blacklist = c("Cortellis_condition", "Cortellis_indication", "ICD9", "ICD10"),
  whitelist = NA
)
}
\arguments{
\item{disNet}{the disease network to cluster and split}

\item{clusterOn}{clustering taking either or both descendants and cross-references (values = "children","xref)}

\item{forwardAmbiguity}{level of forward ambiguity allowed
(default: 10000 ==> ~no filter)}

\item{backwardAmbiguity}{level of backward ambiguity allowed
(default: 1 ==> no ambiguity allowed)}

\item{blacklist}{a vector of databases to avoid when extending (default = NA)}

\item{whitelist}{a vector of databases to trust when extending, only going through these database nodes (default = NA)}
}
\value{
A list of normalized disease networks: setDisNet object
}
\description{
Cluster a disNet object based on cross-references and/or parend/child information.
}
\details{
Cluster a disNet way generate a \emph{setDisNet} object. Based on the \emph{igraph} package to identify
connected components in a graph. These connected identifiers can identified using
cross-reference or hierarchical edges.
}
\examples{
setDisNet <- clusterDisNet(extDisNet,clusterOn = "xref")

}
\seealso{
buildDisNet
}
