% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{connect_to_dodo}
\alias{connect_to_dodo}
\title{Connect to a neo4j DODO database}
\usage{
connect_to_dodo(
  local = F,
  user_name = "",
  password = "",
  local_port = "7474",
  azure_authorization = NULL,
  verbose = T
)
}
\arguments{
\item{local}{Boolean. Connect to a local Neo4J instance, containing the Knowledge Graph. Default: FALSE.}

\item{user_name}{String. The user name to connect to the Graph.}

\item{password}{String. The password to connect to the Graph.}

\item{local_port}{String. Which port is exposed to the local https:// connection. Default: 7474}

\item{azure_authorization}{Object. Location of the rds object with the azure token.}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
Established a connection to local or remote DODO instance
}
\description{
Connect to a neo4j DODO database
}
