% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_concept.R
\name{convert_concept}
\alias{convert_concept}
\title{Convert concept identifiers}
\usage{
convert_concept(
  from,
  to = NULL,
  from.concept = c("Disease", "Phenotype"),
  to.concept = c("Disease", "Phenotype"),
  BA.transitive = 1,
  BA.intransitive = NULL,
  verbose = FALSE
)
}
\arguments{
\item{from}{a vector with identifier to convert formatted as DB:id (eg. "MONDO:0005027)}

\item{to}{database to convert to (default = NULL, no filtering)}

\item{from.concept}{concept (disease or phenotype) of from}

\item{to.concept}{concept (disease or phenotype) to convert to}

\item{verbose}{show query input (default: FALSE)}

\item{FA}{forward ambiguity (default: no filter)}

\item{BA}{backward ambiguity (default: BA = 1)}

\item{FA.transitivity}{forward ambiguity while using transitivity to identify cross-references (default: no filter)}

\item{BA.transitivity}{backward ambiguity while using transitivity to identify cross-references (default: 1)}

\item{FA.non_transitivity}{forward ambiguity while using transitivity to identify cross-references (default: no filter)}

\item{BA.non_transitivity}{backward ambiguity while using transitivity to identify cross-references (default: no filter)}
}
\value{
a dataframe with three columns:
\itemize{
\item from: identifier to convert
\item to: returned conversion
\item hops: length of the shortest path between them
}
}
\description{
Another core functionality is the ability to convert concept identifiers between different
databases and types (disease or phenotype). By default a backward ambiguity of one is taken,
with no limitation on forward ambiguity.
}
