% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dodo.R
\name{convert_dodo}
\alias{convert_dodo}
\title{Convert basic sightings spreadsheet into all \code{dodo} formats}
\usage{
convert_dodo(
  x,
  init.time,
  test.time = as.numeric(format(Sys.Date(), "\%Y")),
  threshold = 0.9,
  unique = TRUE,
  certainty,
  time
)
}
\arguments{
\item{x}{sightings spreadsheet to be converted. Must be a \code{data.frame} with
all of the requisite columns included. See Table S1 in Buchanan et al. (2025)
for an example.}

\item{init.time}{start of the observation period.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}

\item{threshold}{cutoff certainty value for which sightings to consider
certain (for binary certain/uncertain models). Defaults to \code{0.9}.}

\item{unique}{whether to deduplicate the dataset (to mitigate nonidependence
issues). Defaults to \code{TRUE}, in which case the dataset is grouped by every
column except \code{time}, with the most recent row kept for each group.}

\item{certainty}{name of the column with certainty values}

\item{time}{name of the column with time values}
}
\value{
a \code{list} object with 10 elements: the original parameters
\code{init.time}, \code{test.time}, \code{threshold}, and \code{unique}, and the 6 converted
datasets.
}
\description{
Converts a sightings spreadsheet into all the data formats used in \code{dodo}
models.
}
\section{Data Formats}{

\strong{\code{ccon}}: \emph{continuous certain sightings}. The simplest format, a vector of
all the (certain) sighting times of the taxon of interest. See e.g.
\code{\link{SO93F1}}.

\strong{\code{cbin}}: \emph{binary certain sightings}. A vector of 0/1 (certain) sighting
states at all discrete (integer) time intervals between \code{init.time} and
\code{test.time} for the taxon of interest. Similar to \code{ccon}, but note that
multiple records in an interval collapse to 1. See e.g. \code{\link{CB14B1}}.

\strong{\code{cdis}}: \emph{discrete certain sightings}. A vector of (certain) sighting
counts at all discrete (integer) time intervals between \code{init.time} and
\code{test.time} for the taxon of interest. Similar to \code{cbin}, but note that
multiple records in an interval \emph{do not} collapse to 1. See e.g.
\code{\link{BU95F1}}.

\strong{\code{ucon}}: \emph{continuous uncertain sightings}. A \code{data.frame} of all the
sighting times of the taxon of interest, with certainty scores between 0 and
1 associated with each record. See e.g. \code{\link{JR14F1}}.

\strong{\code{ubin}}: \emph{binary uncertain sightings}. A \code{data.frame} with two columns,
the first being the vector of 0/1 certain sighting states at all discrete
(integer) time intervals between \code{init.time} and \code{test.time} for the taxon
of interest, and the second being the equivalent vector for uncertain
sightings. See e.g. \code{\link{KO20B2}}.

\strong{\code{umcd}}: \emph{multi-class discrete uncertain sightings}. A \code{data.frame} of
sighting counts at all discrete (integer) time intervals between \code{init.time}
and \code{test.time} for the taxon of interest. Sightings are split up into
classes based on their \code{certainty} score. See e.g. \code{\link{TH13B1}}.
}

\examples{
# Convert the raw Slender-billed Curlew data
convert_dodo(curlew_raw, init.time = 1817, test.time = 2022, threshold = 0.9,
certainty = "p_ci", time = "year")

}
