% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{describe_concept}
\alias{describe_concept}
\title{Get concept description}
\usage{
describe_concept(ids, sep = ":", verbose = FALSE)
}
\arguments{
\item{ids}{a character vector of  concept identifier to search (e.g. "MONDO:0005027")}

\item{sep}{String. Separator used, default: ":".}

\item{verbose}{return query content (default = FALSE)}
}
\value{
data.table with id, label, definition
}
\description{
Describing provided concept identifiers (when available) using disease label
}
