% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodo-package.R
\docType{package}
\name{dodo-package}
\alias{dodo-package}
\title{dodo: Extinction Models in R}
\description{
A selection of published extinction models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://louis-backstrom.github.io/dodo/}
}

}
\author{
\strong{Maintainer}: Louis Backstrom \email{ljb38@st-andrews.ac.uk}

}
\keyword{internal}
