% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodo.R
\docType{package}
\name{dodo}
\alias{dodo}
\title{dodo: Extinction Models in R}
\description{
dodo implements a selection of published extinction models.
}
\section{Model Functions}{

\itemize{
\item \code{\link[=SL88F1]{SL88F1()}} Springer & Lilje's (1988) "Broken Stick" model.
\item \code{\link[=SS89F1]{SS89F1()}} Strauss & Sadler's (1989) "Classical" model.
\item \code{\link[=SS89B1]{SS89B1()}} Strauss & Sadler's (1989) "Bayesian" model.
\item \code{\link[=SO93F1]{SO93F1()}} Solow's (1993) "Classical" model.
\item \code{\link[=SO93B1]{SO93B1()}} Solow's (1993) "Bayesian" model.
\item \code{\link[=SO93F2]{SO93F2()}} Solow's (1993) "Declining" model.
\item \code{\link[=MA94F1]{MA94F1()}} Marshall's (1994) "Distribution-free" model.
\item \code{\link[=BU95F1]{BU95F1()}} Burgman et al.'s (1995) "Discrete-time" model.
\item \code{\link[=BU95F2]{BU95F2()}} Burgman et al.'s (1995) "Runs Test" model.
\item \code{\link[=BU95F3]{BU95F3()}} Burgman et al.'s (1995) "Empty Cells" model.
\item \code{\link[=MA97F1]{MA97F1()}} Marshall's (1997) "Non-random" model.
\item \code{\link[=MC98F1]{MC98F1()}} McCarthy's (1998) "Partial Solow" model.
\item \code{\link[=MC99F1]{MC99F1()}} McFarlane's (1999) "Median Gap" model.
\item \code{\link[=WM99B1]{WM99B1()}} Weiss & Marshall's (1999) "Discrete-time" model.
\item \code{\link[=RS03F1]{RS03F1()}} Roberts & Solow's (2003) "Optimal Linear Estimation" model.
\item \code{\link[=SR03F1]{SR03F1()}} Solow & Roberts' (2003) "Non-parametric" model.
\item \code{\link[=MC06F1]{MC06F1()}} McInerny et al.'s (2006) "Sighting Rate" model.
\item \code{\link[=RO09B1]{RO09B1()}} Rout et al.'s (2009) "Declining" model.
\item \code{\link[=JE10F1]{JE10F1()}} Jarić & Ebenhard's (2010) "Stationary" model.
\item \code{\link[=JE10F2]{JE10F2()}} Jarić & Ebenhard's (2010) "Non-stationary" model.
\item \code{\link[=BR12F1]{BR12F1()}} Bradshaw et al.'s (2012) "Inverse-weighted McInerny" model.
\item \code{\link[=SO12B1]{SO12B1()}} Solow et al.'s (2012) "Uncertain" model.
\item \code{\link[=TH13B1]{TH13B1()}} Thompson et al.'s (2013) "Deterministic" model.
\item \code{\link[=TH13B2]{TH13B2()}} Thompson et al.'s (2013) "Quenched" model.
\item \code{\link[=TH13B3]{TH13B3()}} Thompson et al.'s (2013) "Annealed" model.
\item \code{\link[=AL14B1]{AL14B1()}} Alroy's (2015) "Creeping Shadow of a Doubt" model.
\item \code{\link[=CB14B1]{CB14B1()}} Caley & Barry's (2014) "Constant" model.
\item \code{\link[=CB14B2]{CB14B2()}} Caley & Barry's (2014) "Non-constant" model.
\item \code{\link[=JR14F1]{JR14F1()}} Jarić & Roberts' (2014) "Solow" model.
\item \code{\link[=LE14B1]{LE14B1()}} Lee's (2014) "Simple" model.
\item \code{\link[=LE14B2]{LE14B2()}} Lee et al.'s (2014) "Variable Reliability" model.
\item \code{\link[=SB14B1]{SB14B1()}} Solow & Beet's (2014) "Model 1" model.
\item \code{\link[=SB14B2]{SB14B2()}} Solow & Beet's (2014) "Model 2" model.
\item \code{\link[=AL15B1]{AL15B1()}} Alroy's (2015) "Agnostic" model.
\item \code{\link[=SO16B1]{SO16B1()}} Solow's (2016) "Retrospective" model.
\item \code{\link[=SO16B2]{SO16B2()}} Solow's (2016) "Sequential" model.
\item \code{\link[=WA16B1]{WA16B1()}}	Wang et al.'s (2016) "Adaptive" model.
\item \code{\link[=TH17I1]{TH17I1()}} Thompson et al.'s (2017) "Iterative" model.
\item \code{\link[=BR19F1]{BR19F1()}} Brook et al.'s (2019) "Solow" model.
\item \code{\link[=BR19F2]{BR19F2()}} Brook et al.'s (2019) "Solow & Roberts" model.
\item \code{\link[=BR19F3]{BR19F3()}} Brook et al.'s (2019) "Roberts & Solow" model.
\item \code{\link[=BR19F4]{BR19F4()}} Brook et al.'s (2019) "McInerny" model.
\item \code{\link[=BR19F5]{BR19F5()}} Brook et al.'s (2019) "Last Appearance Date" model.
\item \code{\link[=TH19B1]{TH19B1()}} Thompson et al.'s (2017) "Bayesian Updating" model.
\item \code{\link[=KO20B1]{KO20B1()}} Kodikara et al.'s (2020) "Certain-only" model.
\item \code{\link[=KO20B2]{KO20B2()}} Kodikara et al.'s (2020) "Uncertain" model.
\item \code{\link[=KO21B1]{KO21B1()}} Kodikara et al.'s (2021) "Change-point" model.
\item \code{\link[=BA26B1]{BA26B1()}} Backstrom et al.'s (2026) "Certain-only No-effort" model.
\item \code{\link[=BA26B2]{BA26B2()}} Backstrom et al.'s (2026) "Uncertain No-effort" model.
\item \code{\link[=BA26B3]{BA26B3()}} Backstrom et al.'s (2026) "Certain-only Effort" model.
}
}

\section{Other Functions}{

\itemize{
\item \code{\link[=convert_dodo]{convert_dodo()}} Convert basic sightings spreadsheet into all \code{dodo}
formats.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://louis-backstrom.github.io/dodo/}
}

}
\author{
\strong{Maintainer}: Louis Backstrom \email{ljb38@st-andrews.ac.uk}

}
