% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dodo}
\alias{dodo}
\title{Dodo Sightings}
\format{
A numeric vector with 10 values.
}
\usage{
dodo
}
\description{
Dodo (\emph{Raphus cucullatus}) sighting records from Roberts & Solow (2003).
}
\references{
\strong{Key Reference}

Roberts, D. L., & Solow, A. R. (2003). Flightless birds: when did the dodo
become extinct? \emph{Nature}, 426(6964), 245. \doi{10.1038/426245a}
}
\keyword{datasets}
