% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_disNet.R
\name{filter_by_database}
\alias{filter_by_database}
\title{Filter a disNet by database}
\usage{
filter_by_database(disNet, databases)
}
\arguments{
\item{disNet}{a disease network}

\item{databases}{a character vector of databases of interest (format: c("MONDO","EFO"))}
}
\value{
A normalized disease network
}
\description{
Filter a disNet to keep only identifiers of databases of interest
}
\examples{
disNet <- build_disNet(term = "epilep")
filtDisNet <- filter_by_database(disNet, databases = "EFO")

}
