% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_disNet.R
\name{filter_by_id}
\alias{filter_by_id}
\title{Filter a disNet by identifiers}
\usage{
filter_by_id(disNet, diseaseID)
}
\arguments{
\item{disNet}{a disease network}

\item{diseaseID}{a character vector of disease ids of interest (format: "MONDO:0005027")}
}
\value{
A normalized disease network

filtId <- disNet1$nodes$id
disNet2 <- buildDisNetByTerm("epilep",fields = c("synonym","label","definition"))
filtDisNet <- filtDisNetByID(disNet2, filtId)
}
\description{
Filtering a disNet to keep only diseases of interest
}
