% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_concepts.R
\name{find_id}
\alias{find_id}
\title{Find concept by identifier}
\usage{
find_id(ids)
}
\arguments{
\item{ids}{a character vector of concept to search formatted as "DB:id" (e.g. "MONDO:0005027")}
}
\value{
A character vector of concept identfiers in DODO
}
\description{
Find concept by identifier
}
\examples{
toRet <- find_id(ids = c("MONDO:0005027","MONDO:0100033"))

}
