% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fox}
\alias{fox}
\title{Tasmanian Fox Carcass Recoveries}
\format{
A numeric vector with 4 values.
}
\usage{
fox
}
\description{
European Red Fox (\emph{Vulpes vulpes}) carcass recovery dates from Caley & Barry
(2014).
}
\references{
\strong{Key Reference}

Caley, P., & Barry, S. C. (2014). Quantifying extinction probabilities from
sighting records: inference and uncertainties. \emph{PLoS One}, 9(4), e95857.
\doi{10.1371/journal.pone.0095857}
}
\keyword{datasets}
