% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gerygone}
\alias{gerygone}
\title{Lord Howe Gerygone Sightings}
\format{
a \code{data.frame} with 3 columns and 236 rows.
}
\usage{
gerygone
}
\description{
Frequency-based collection-effort sighting record of the Lord Howe Gerygone
(\emph{Gerygone insularis}) based on data from the Atlas of Living Australia.
}
\note{
Includes all observations of birds within 10km of Lord Howe Island with
valid spatial coordinates. The index of collection effort is the total
number of birds recorded in that year.
}
\references{
\strong{Key Reference}
Atlas of Living Australia occurrence download, accessed on 25 August 2025.
\doi{10.26197/ala.dad3870b-1b79-4600-ad12-af429feb1ad2}

\strong{Other References}
McCarthy, M. A. (1998). Identifying declining and threatened species with
museum data. \emph{Biological Conservation}, 83(1), 9-17.
\doi{10.1016/S0006-3207(97)00048-7}
}
\keyword{datasets}
