% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ontology.R
\name{get_ontology}
\alias{get_ontology}
\title{Return entire disease ontology}
\usage{
get_ontology(database)
}
\arguments{
\item{database}{disease ontology to return.}
}
\description{
Return all nodes for a particular disease ontology
}
\seealso{
\code{\link{list_database}}
}
