% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_concept.R
\name{get_related}
\alias{get_related}
\title{Get related concepts}
\usage{
get_related(
  from,
  to = NULL,
  from.concept = c("Disease", "Phenotype"),
  to.concept = c("Disease", "Phenotype"),
  deprecated = FALSE,
  transitive_ambiguity = 1,
  intransitive_ambiguity = NULL,
  verbose = FALSE
)
}
\arguments{
\item{from}{a vector with identifier to convert formatted as DB:id (eg. "MONDO:0005027)}

\item{to}{database to convert to (default = NULL, no filtering)}

\item{from.concept}{concept (disease or phenotype) of from}

\item{to.concept}{concept (disease or phenotype) to convert to}

\item{deprecated}{include deprecated identifiers (default: false)}

\item{transitive_ambiguity}{backward ambiguity while using transitivity to identify cross-references (default: 1)}

\item{intransitive_ambiguity}{specification for backward ambiguity used in the final step of conversion (default: no filter)}

\item{verbose}{show query input (default: FALSE)
#'}
}
\value{
a dataframe with three columns:
\itemize{
\item from: identifier to convert
\item to: returned conversion
}
}
\description{
The specific conversion procedure is recommended for the ontologies that
consider disease concepts that are less connected through \emph{is_xref} edges
(e.g. ICD10, ICD9, ClinVar) and are mostly only related to other identifiers
using \emph{is_related} edges. This limits or removes the effect of the transitive
mapping and not return all cross-references. Therefore, for these ontologies it
is recommend that in addition to the standard conversion, an additional step
of intransitive mapping is performed which is implemented in the function \emph{get_related}.
}
