% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JE10F1.R
\name{JE10F1}
\alias{JE10F1}
\title{Jarić & Ebenhard's (2010) "Stationary" model}
\usage{
JE10F1(
  records,
  init.time = min(records),
  test.time = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{records}{sighting records in \code{ccon} format (see
\code{\link{convert_dodo}} for details).}

\item{init.time}{start of the observation period. Defaults to the time of
the first sighting, in which case this sighting is removed from the record.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}
}
\value{
a \code{list} object with the original parameters and the p-value
included as elements.
}
\description{
Equation 4 from Jarić & Ebenhard 2010. Estimates a p-value for testing
competing hypotheses of extinction/non-extinction.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\examples{
# Run the Black-footed Ferret analysis from Jarić & Ebenhard 2010
JE10F1(ferret$ccon - 5, test.time = 223) # shift dates to align with paper
# Run an example analysis using the Slender-billed Curlew data
\dontrun{
JE10F1(curlew$ccon, init.time = 1817, test.time = 2022)
}

}
\references{
\strong{Key Reference}

Jarić, I., & Ebenhard, T. (2010). A method for inferring extinction based on
sighting records that change in frequency over time. \emph{Wildlife Biology},
16(3), 267-275. \doi{10.2981/09-044}
}
\seealso{
\code{\link[=JE10F2]{JE10F2()}}
}
