% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JE10F1.R
\name{JE10F1}
\alias{JE10F1}
\title{Jarić & Ebenhard's (2010) "Stationary" model}
\usage{
JE10F1(records, test.time = as.numeric(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{records}{numeric vector object containing all sighting records of the
taxon of interest.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}
}
\value{
a \code{list} object with the original parameters and the p-value
included as elements.
}
\description{
Equation 4 from Jarić & Ebenhard 2010. Estimates a p-value for testing
competing hypotheses of extinction/non-extinction.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\examples{
# Run the Black-footed Ferret analysis from Jarić & Ebenhard 2010
JE10F1(ferret1 - 5, test.time = 223) # shift dates to align with paper

}
\references{
\strong{Key Reference}

Jarić, I., & Ebenhard, T. (2010). A method for inferring extinction based on
sighting records that change in frequency over time. \emph{Wildlife Biology},
16(3), 267-275. \doi{10.2981/09-044}
}
\seealso{
\code{\link[=JE10F2]{JE10F2()}}
}
