% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KO21B1.R
\name{KO21B1}
\alias{KO21B1}
\title{Kodikara et al.'s (2021) "Change-point" Model}
\usage{
KO21B1(
  records,
  alpha = 0.05,
  init.time = min(records$time),
  test.time = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{records}{\code{data.frame} with two columns: \code{time} and \code{certain}. The
\code{time} column contains the date of all sightings, which are each either
certain (\code{certain = TRUE}) or uncertain (\code{certain = FALSE}).}

\item{alpha}{desired threshold level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} credible interval.}

\item{init.time}{start of the observation period. Defaults to the time of
the first sighting, in which case this sighting is removed from the record.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}
}
\value{
a \code{list} object with the original parameters and the p(extant),
point estimate, and credible interval included as elements. The credible
interval is a two-element numeric vector called \code{cred.int}.
}
\description{
The model from Kodikara et al. 2021. Estimates a posterior probability that
the species is extant at the test time, and a point estimate and one-tailed
\eqn{1 - \alpha} credibile interval on the time of extinction.
}
\note{
Sampling effort is assumed to be constant.
}
\references{
\strong{Key Reference}

Kodikara, S., Demirhan, H., Wang, Y., Solow, A., & Stone, L. (2020).
Inferring extinction year using a Bayesian approach. \emph{Methods in Ecology}
\emph{and Evolution}, 11(8), 964-973. \doi{10.1111/2041-210x.13408}
}
