% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-utilities.R
\name{load_concept_names}
\alias{load_concept_names}
\title{Feeding DODO: load concept names in DODO DB}
\usage{
load_concept_names(toImport, concept)
}
\arguments{
\item{toImport}{a data.frame with the following columns:
\itemize{
\item database: \emph{character} name of the database (mandatory)
\item origin: \emph{character} name of the database from which the concept was
taken (optional). If not provided or NA, the database field is used.
\item shortID: \emph{character} short concept ID (mandatory)
}}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
