% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC98F1.R
\name{MC98F1}
\alias{MC98F1}
\title{McCarthy's (1998) "Partial Solow" model}
\usage{
MC98F1(records)
}
\arguments{
\item{records}{\code{data.frame} with three columns: \code{time}, \code{records}, and
\code{effort}. The \code{time} column must extend from the start of the observation
period (which may be prior to the first sighting) to the end (typically the
present day), with evenly-spaced temporal intervals (typically years).}
}
\value{
a \code{list} object with the original parameters and the p-value
included as elements.
}
\description{
Equation 2 from McCarthy 1998. Estimates a p-value for testing
competing hypotheses of extinction/non-extinction.
}
\note{
All sighting records are assumed to be certain.
}
\references{
\strong{Key Reference}

McCarthy, M. A. (1998). Identifying declining and threatened species with
museum data. \emph{Biological Conservation}, 83(1), 9-17.
\doi{10.1016/S0006-3207(97)00048-7}
}
