% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC99F1.R
\name{MC99F1}
\alias{MC99F1}
\title{McFarlane's (1999) "Median Gap" model}
\usage{
MC99F1(
  records,
  alpha = 0.05,
  test.time = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{records}{sighting records in \code{ccon} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}
}
\value{
a \code{list} object with the original parameters and the p-value and
confidence interval included as elements. The confidence interval is
a two-element numeric vector called \code{conf.int}.
}
\description{
Equation 3 from McFarlane 1999. Estimates a p-value for testing competing
hypotheses of extinction/non-extinction, and a one-tailed \eqn{1 - \alpha}
confidence interval on the time of extinction.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\examples{
# Run an example analysis using the Woolly Mammoth data
MC99F1(mammoth, test.time = -11000)
# Run an example analysis using the Slender-billed Curlew data
MC99F1(curlew$ccon, test.time = 2022)

}
\references{
\strong{Key Reference}

McFarlane, D. A. (1999). A Comparison of Methods for the Probabilistic
Determination of Vertebrate Extinction Chronologies. In R. D. E. MacPhee
(Ed.), \emph{Extinctions in Near Time} (pp. 95-103). Springer US.
\doi{10.1007/978-1-4757-5202-1_5}
}
