% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monk_seal}
\alias{monk_seal}
\title{Caribbean Monk Seal sightings}
\format{
A numeric vector with 5 values.
}
\usage{
monk_seal
}
\description{
Caribbean Monk Seal (\emph{Monachus tropicalis}) sighting records from Solow
(1993) based on data presented in LeBouef et al. (1986).
}
\references{
\strong{Key Reference}

Solow, A. R. (1993). Inferring Extinction from Sighting Data. \emph{Ecology},
74(3), 962-964. \doi{10.2307/1940821}

\strong{Other References}

LeBoeuf, B. J., Kenyon, K. W., & Villa‐Ramirez, B. (1986). The Caribbean monk
seal is extinct. \emph{Marine Mammal Science}, 2(1), 70-72.
\doi{10.1111/j.1748-7692.1986.tb00028.x}
}
\keyword{datasets}
