% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-utilities.R
\name{plot.disNet}
\alias{plot.disNet}
\title{Visualize a disease network}
\usage{
\method{plot}{disNet}(
  disNet,
  relations = c("xref", "child", "parent"),
  number = 500,
  igraph = T,
  ...
)
}
\arguments{
\item{disNet}{a disease network}

\item{relations}{relationships to plot, can take "xref","child" and/or "parent" as values}

\item{number}{Number of nodes to plot without igraph. If the number of nodes exceeds, igraph will be used (integer)}

\item{igraph}{use visIgraphLayout() (default = TRUE)}
}
\value{
A \code{\link{visNetwork}} object.
}
\description{
Visualize a disNet or setDisNet object as a visNetwork plot. It can be specified to show only a particular edge type
and by default only 500 disease nodes are plotted.
}
\examples{
disNet <- extendDisNet(buildDisNetByID(id = "MONDO:0005027"))
plot.disNet(disNet) 

}
