% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-utilities.R
\name{plot.setDisNet}
\alias{plot.setDisNet}
\title{Visualize one cluster of set disease networks}
\usage{
\method{plot}{setDisNet}(
  disNet,
  relations = c("xref", "child", "parent"),
  cluster = NULL,
  number = 500,
  igraph = T,
  labelling = FALSE,
  ...
)
}
\arguments{
\item{disNet}{a disease network}

\item{relations}{relationships to plot, can take "xref","child" and/or "parent" as values}

\item{cluster}{cluster to plot}

\item{number}{Number of nodes to plot without igraph. If the number of nodes exceeds, igraph will be used (integer)}

\item{igraph}{use visIgraphLayout() (default = TRUE)}

\item{labelling}{use either disease identifier or both disease identifier and label
to label nodes (default = FALSE)}
}
\value{
A \code{\link{visNetwork}} object.
}
\description{
Visualize one cluster of a setDisNet object as a visNetwork plot. It can be specified to show only a particular edge type
and by default only 500 disease nodes are plotted.
}
\examples{
disNet <- clusterDisNet(extendDisNet(buildDisNetByTerm(id = "ALS")), clusterOn = "xref")#' 
plot(disNet, cluster = 1) 

}
