% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-utilities.R
\name{plot_relations}
\alias{plot_relations}
\title{Identify and show all relationships}
\usage{
plot_relations(
  df,
  step = NULL,
  transitive.ambiguity = 1,
  intransitive.ambiguity = NULL
)
}
\arguments{
\item{df}{Dataframe. Dataframe with columns dbid1, type, dbid2, confidence}
}
\description{
Takes the dataframe that was obtained from \link{convert_concept} as input and will query all cross-reference edges.
For use of forward and backward ambiguity as well as black- and whitelist, see extendDisNet.
}
\details{
The initial identifier is represented by a triangle, each node is colored based on the database. The edges
between nodes are colored based on the ambiguity equal to one or larger.

The edges are formatted based on forward ambiguity and type of the edge.
The colour of the edges is based on ambiguity equal to one or larger than one.
This information is encoded when hovering over the edge.
The solid lines are used for \emph{is_xref} edges, while dashed lines are
used from \emph{is_related} edges.
}
