% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RO09B1.R
\name{RO09B1}
\alias{RO09B1}
\title{Rout et al.'s (2009) "Declining" model}
\usage{
RO09B1(records, pi = 0.5, n.chains = 4, n.iter = 110000, n.burnin = 10000)
}
\arguments{
\item{records}{sighting records in \code{cbin} format (see
\code{\link{convert_dodo}} for details).}

\item{pi}{prior probability that \eqn{H_0} is true (defaults to
\eqn{\pi = 0.5}).}

\item{n.chains}{number of MCMC chains to run. Defaults to 4.}

\item{n.iter}{number of iterations in each chain. Defaults to 110,000.}

\item{n.burnin}{number of iterations to discard as burn-in. Defaults to
10,000.}
}
\value{
a \code{list} object with the original parameters and the rate of
decline (\eqn{a}), the Bayes factor, and p(extant) included as elements.
}
\description{
Equation 6 and others from Rout et al. 2009. Estimates a pre-extinction rate
of decline, a Bayes factor comparing competing hypotheses of extinction
/ persistence, and a posterior probability that the species is extant at the
end of the observation period.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant. The Bayes Factor presented here is the inverse of the Bayes
Factor as presented in the original paper, to allow for comparability with
other models in this package (values > 1 imply extinction).
}
\examples{
\dontrun{
# Run the bitterweed analysis from Rout et al. 2009
RO09B1(bitterweed)
# Run an example analysis using the Slender-billed Curlew data
RO09B1(curlew$cbin)
}

}
\references{
\strong{Key Reference}

Rout, T. M., Salomon, Y., & McCarthy, M. A. (2009). Using sighting records
to declare eradication of an invasive species. \emph{Journal of Applied Ecology},
46(1), 110-117. \doi{10.1111/j.1365-2664.2008.01586.x}

\strong{Other References}

Solow, A. R. (1993). Inferring Extinction from Sighting Data. \emph{Ecology},
74(3), 962-964. \doi{10.2307/1940821}
}
