% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS03F1.R
\name{RS03F1}
\alias{RS03F1}
\title{Roberts & Solow's (2003) "Optimal Linear Estimation" model}
\usage{
RS03F1(
  records,
  alpha = 0.05,
  k = 10,
  test.time = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{records}{numeric vector object containing all sighting records of the
taxon of interest.}

\item{alpha}{desired significance level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} confidence interval.}

\item{k}{number of most recent sighting records to use (defaults to 5).}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}
}
\value{
a \code{list} object with the original parameters and the p-value, point
estimate, and confidence interval included as elements. The confidence
interval is a two-element numeric vector called \code{conf.int}.
}
\description{
Equations from Roberts & Solow 2003 (Equations 12-18 from Solow 2005).
Estimates a p-value for testing competing hypotheses of
extinction/non-extinction, and a two-tailed \eqn{1 - \alpha} confidence
interval and point estimate on the time of extinction.
}
\note{
All sighting records are assumed to be certain.
}
\references{
\strong{Key Reference}

Roberts, D. L., & Solow, A. R. (2003). Flightless birds: when did the dodo
become extinct? \emph{Nature}, 426(6964), 245. \doi{10.1038/426245a}

\strong{Other References}

Solow, A. R. (2005). Inferring extinction from a sighting record.
\emph{Mathematical Biosciences}, 195(1), 47-55. \doi{10.1016/j.mbs.2005.02.001}

Rivadeneira, M. M., Hunt, G., & Roy, K. (2009). The use of sighting records
to infer species extinctions: an evaluation of different methods. \emph{Ecology},
90(5), 1291-1300. \doi{10.1890/08-0316.1}
}
