% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_concept.R
\name{searchDODO}
\alias{searchDODO}
\title{Search concept}
\usage{
searchDODO(
  searchTerm,
  database = NULL,
  type = c(),
  fields = c("label", "synonym"),
  exact = FALSE
)
}
\arguments{
\item{searchTerm}{a term to search for, can be either an ID or a term}

\item{database}{if provided,}

\item{type}{the type of the searchTerm, either id or name}

\item{fields}{If a term is search, fields can be specified to label and synonym. By default all fields are used.}

\item{exact}{exact match term}
}
\description{
Search concept
}
