% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-utilities.R
\name{show_relations}
\alias{show_relations}
\title{Identify and show all relationships}
\usage{
show_relations(
  id,
  step = NULL,
  transitive.ambiguity = 1,
  intransitive.ambiguity = NULL
)
}
\arguments{
\item{step}{the depth of the search (default: 10000 ==> ~exhaustive)}

\item{transitive.ambiguity}{backward ambiguity while using transitivity to identify cross-references (default: 1)}

\item{intransitive.ambiguity}{backward ambiguity while using transitivity to identify cross-references (default: no filter)}

\item{dbid}{concept identifier formatted as db:id}

\item{relations}{the kind of relationships to take into account -> xref}

\item{blacklist}{a vector of databases to avoid when extending (default = NA)}

\item{whitelist}{a vector of databases to trust when extending, only going through these database nodes (default = NA)}
}
\description{
Takes a identifier (full length, e.g. "MONDO:0005027") as input and will query all cross-reference edges.
For use of forward and backward ambiguity as well as black- and whitelist, see extendDisNet.
}
\details{
The initial identifier is represented by a triangle, each node is colored based on the database. The edges
between nodes are colored based on the ambiguity equal to one or larger.
}
\seealso{
extendDisNet
}
