% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO12B1.R
\name{SO12B1}
\alias{SO12B1}
\title{Solow et al.'s (2012) "Uncertain" model}
\usage{
SO12B1(records, init.time, increment = 0.01, gamma = 1, record.check = "error")
}
\arguments{
\item{records}{sighting records in \code{ubin} format (see
\code{\link{convert_dodo}} for details). Note that all uncertain sightings
must follow the final certain sighting. If any uncertain sightings occur
prior to the final certain sighting, these are handled as specified by
\verb{record check}.}

\item{init.time}{start of the observation period.}

\item{increment}{step size used for integration. Defaults to 0.01, following
the original code from Solow & Beet 2014.}

\item{gamma}{parameter for the exponential prior. Defaults to 1.}

\item{record.check}{what to do with uncertain sightings prior to the final
certain sighting. Three options: \code{error} (default, throws an error),
\code{remove} (problematic sightings are removed from the data), and \code{coerce}
(problematic sightings are treated as certain sightings).}
}
\value{
a \code{list} object with the original parameters and the Bayes factor
included as elements.
}
\description{
The model from Solow et al. 2012. Estimates a Bayes factor comparing
competing hypotheses of extinction / persistence.
}
\note{
Sampling effort is assumed to be constant. Uses an exponential prior.
}
\examples{
# Run the Ivory-billed Woodpecker analysis from Solow et al. 2012
SO12B1(
  records = woodpecker$ubin, init.time = 1897, increment = 0.01,
  record.check = "coerce"
)
# Run an example analysis using the Slender-billed Curlew data
\dontrun{
SO12B1(curlew$ubin, init.time = 1817, increment = 0.01)
}

}
\references{
\strong{Key Reference}

Solow, A., Smith, W., Burgman, M., Rout, T., Wintle, B., & Roberts, D.
(2012). Uncertain sightings and the extinction of the Ivory-billed
Woodpecker. \emph{Conservation Biology}, 26(1), 180-184.
\doi{10.1111/j.1523-1739.2011.01743.x}

\strong{Other References}

Solow, A. R., & Beet, A. R. (2014). On uncertain sightings and inference
about extinction. \emph{Conservation Biology}, 28(4), 1119-1123.
\doi{10.1111/cobi.12309}
}
