% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO16B2.R
\name{SO16B2}
\alias{SO16B2}
\title{Solow's (2016) "Sequential" model}
\usage{
SO16B2(records, init.time = NULL, test.time = NULL, curr.time = NULL)
}
\arguments{
\item{records}{sighting records in \code{cbin} format (see
\code{\link{convert_dodo}} for details).}

\item{init.time}{start of the observation period. Defaults to \code{NULL}; this
parameter is only necessary if specifying a different \code{test.time} to the end
of the observation period.}

\item{test.time}{time point to retrospectively calculate extinction
probability at. Defaults to \code{NULL}, in which case the probability is
estimated for the end of the observation period.}

\item{curr.time}{end of the observation period. Defaults to \code{NULL}; this
parameter is only necessary if specifying a different \code{test.time} to the end
of the observation period. If specified, \code{curr.time} - \code{init.time} must equal
\code{length(records)}.}
}
\value{
a \code{list} object with the original parameters and p(extant) included
as elements.
}
\description{
Equation 9 from Solow 2016. Estimates a posterior probability that the
species is extant.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant. Uses the uniform prior (1 / (m + T)) specified on p. 797 of
Solow (2016).
}
\examples{
# Run the Dodo analysis from Solow 2016
SO16B2(
  records = as.integer((min(dodos) + 1):2015 \%in\% dodos),
  init.time = min(dodos), test.time = 1672, curr.time = 2015
)
# NB: Solow 2016 presents p(extinct), which is 1 - p(extant) used here.
\dontrun{
# Run an example analysis using the Slender-billed Curlew data
SO16B2(curlew$cbin)
}

}
\references{
\strong{Key Reference}

Solow, A. R. (2016). A simple Bayesian method of inferring extinction:
comment. \emph{Ecology}, 97(3), 796-798. \doi{10.1890/15-0336.1}

\strong{Other References}

Alroy, J. (2014). A simple Bayesian method of inferring extinction.
\emph{Paleobiology}, 40(4), 584-607. \doi{10.1666/13074}

Alroy, J. (2016). A simple Bayesian method of inferring extinction: reply.
\emph{Ecology}, 97(3), 798-800. \doi{10.1002/ecy.1321}
}
\seealso{
\code{\link[=AL14B1]{AL14B1()}}; \code{\link[=SO16B1]{SO16B1()}}
}
