% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO93B1.R
\name{SO93B1}
\alias{SO93B1}
\title{Solow's (1993) "Bayesian" model}
\usage{
SO93B1(
  records,
  alpha = 0.05,
  init.time = min(records),
  test.time = as.numeric(format(Sys.Date(), "\%Y")),
  prior = runif(n = 1e+06, min = 0, max = 1),
  keep.dists = FALSE
)
}
\arguments{
\item{records}{numeric vector object containing all sighting records of the
taxon of interest.}

\item{alpha}{desired threshold level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} credible interval.}

\item{init.time}{start of the observation period. Defaults to the time of
the first sighting, in which case this sighting is removed from the record.}

\item{test.time}{end of the observation period, typically the present day
(defaults to the current year).}

\item{prior}{prior distribution of extinction probability.
Defaults to a \eqn{U(0, 1)} distribution.}

\item{keep.dists}{whether to include the posterior distribution vectors in
the final function output or not (defaults to \code{FALSE}).}
}
\value{
a \code{list} object with the original parameters and the Bayes factor,
point estimate, and credible interval included as elements. The credible
interval is a two-element numeric vector called \code{cred.int}. The prior and
posterior distribution vectors are also included, if \code{keep.dists} is set
to \code{TRUE}.
}
\description{
Equation 3 and others from Solow 1993. Estimates a Bayes factor and
posterior distribution on probability that the species is extant, with
associated point estimate and two-sided credible interval.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\examples{
# Run the Caribbean Monk Seal analysis from Solow 1993
SO93B1(monk_seal, test.time = 1992)

}
\references{
\strong{Key Reference}

Solow, A. R. (1993). Inferring Extinction from Sighting Data. \emph{Ecology},
74(3), 962-964. \doi{10.2307/1940821}
}
\seealso{
\code{\link[=SO93F1]{SO93F1()}}
}
