% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS89B1.R
\name{SS89B1}
\alias{SS89B1}
\title{Strauss & Sadler's (1989) "Bayesian" model}
\usage{
SS89B1(records, alpha = 0.05, length.out = 1e+07, scale = 0.01)
}
\arguments{
\item{records}{sighting records in \code{ccon} format (see
\code{\link{convert_dodo}} for details).}

\item{alpha}{desired threshold level (defaults to \eqn{\alpha = 0.05}) of
the \eqn{1 - \alpha} credible interval.}

\item{length.out}{number of posterior samples to generate (defaults to 100
thousand).}

\item{scale}{factor to scale sighting records by. Defaults to 0.01; adjust
if warned.}
}
\value{
a \code{list} object with the original parameters and the point estimate
and credible interval included as elements. The credible interval is a
two-element numeric vector called \code{cred.int}.
}
\description{
Equation 26 from Strauss & Sadler 1989, assuming the prior distribution
from equation 22. Estimates a posterior distribution on time of extinction,
with associated point estimate and one-sided credible interval.
}
\note{
All sighting records are assumed to be certain and sampling effort is assumed
to be constant.
}
\examples{
# Run an example analysis using the Caribbean Monk Seal data
SS89B1(monk_seal, length.out = 1e5)
\dontrun{
# Run an example analysis using the Slender-billed Curlew data
SS89B1(curlew$ccon, length.out = 1e5)
}

}
\references{
\strong{Key Reference}

Strauss, D., & Sadler, P. M. (1989). Classical Confidence Intervals and
Bayesian Probability Estimates for Ends of Local Taxon Ranges.
\emph{Mathematical Geology}, 21(4), 411-421. \doi{10.1007/Bf00897326}
}
\seealso{
\code{\link[=SS89F1]{SS89F1()}}
}
